@@ -1,11 +1,12 @@
 module github.com/temporalio/tcld
 
-go 1.18
+go 1.20
 
 require (
 	github.com/go-playground/validator/v10 v10.13.0
 	github.com/gogo/protobuf v1.3.2
 	github.com/golang/mock v1.6.0
+	github.com/golang/protobuf v1.5.2
 	github.com/google/uuid v1.3.0
 	github.com/gosuri/uilive v0.0.4
 	github.com/kylelemons/godebug v1.1.0
@@ -13,6 +14,7 @@ require (
 	github.com/urfave/cli/v2 v2.25.3
 	go.uber.org/fx v1.19.2
 	go.uber.org/multierr v1.6.0
+	golang.org/x/mod v0.12.0
 	google.golang.org/grpc v1.54.0
 )
 
@@ -21,7 +23,6 @@ require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/go-playground/locales v0.14.1 // indirect
 	github.com/go-playground/universal-translator v0.18.1 // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/leodido/go-urn v1.2.3 // indirect
 	github.com/mattn/go-isatty v0.0.19 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
