@@ -10,6 +10,7 @@ import (
 
 	"github.com/temporalio/tcld/protogen/api/auth/v1"
 	"github.com/temporalio/tcld/protogen/api/authservice/v1"
+	"github.com/temporalio/tcld/protogen/api/sink/v1"
 
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/suite"
@@ -36,6 +37,19 @@ type NamespaceTestSuite struct {
 }
 
 func (s *NamespaceTestSuite) SetupTest() {
+	feature, err := NewFeatureCommand()
+	s.Require().NoError(err)
+
+	s.cliApp = &cli.App{
+		Name:     "test",
+		Commands: []*cli.Command{feature.Command},
+		Flags: []cli.Flag{
+			AutoConfirmFlag,
+		},
+	}
+	err = s.RunCmd("feature", "toggle-export")
+	s.Require().NoError(err)
+
 	s.mockCtrl = gomock.NewController(s.T())
 	s.mockService = namespaceservicemock.NewMockNamespaceServiceClient(s.mockCtrl)
 	s.mockReqService = requestservicemock.NewMockRequestServiceClient(s.mockCtrl)
@@ -56,13 +70,7 @@ func (s *NamespaceTestSuite) SetupTest() {
 	out, err := NewNamespaceCommand(getNamespaceClientFn, getRequestClientFn)
 	s.Require().NoError(err)
 	AutoConfirmFlag.Value = true
-	s.cliApp = &cli.App{
-		Name:     "test",
-		Commands: []*cli.Command{out.Command},
-		Flags: []cli.Flag{
-			AutoConfirmFlag,
-		},
-	}
+	s.cliApp.Commands = []*cli.Command{out.Command}
 }
 
 func (s *NamespaceTestSuite) RunCmd(args ...string) error {
@@ -70,6 +78,7 @@ func (s *NamespaceTestSuite) RunCmd(args ...string) error {
 }
 
 func (s *NamespaceTestSuite) AfterTest(suiteName, testName string) {
+	defer os.Remove(getFeatureFlagConfigFilePath())
 	s.mockCtrl.Finish()
 }
 
@@ -1074,6 +1083,108 @@ func (s *NamespaceTestSuite) TestClearCertificateFilters() {
 	}
 }
 
+func (s *NamespaceTestSuite) TestUpdateCodecServer() {
+	ns := "ns1"
+	type morphGetResp func(*namespaceservice.GetNamespaceResponse)
+	type morphUpdateReq func(*namespaceservice.UpdateNamespaceRequest)
+
+	tests := []struct {
+		args         []string
+		expectGet    morphGetResp
+		expectErr    bool
+		expectUpdate morphUpdateReq
+	}{
+		{
+
+			args:      []string{"namespace", "update-codec-server"},
+			expectErr: true,
+		},
+		{
+			args:      []string{"namespace", "update-codec-server", "--namespace", ns},
+			expectErr: true,
+		},
+		{
+			args:      []string{"n", "ucs", "-n", ns, "-endpoint", "fakehost:9999"},
+			expectGet: func(g *namespaceservice.GetNamespaceResponse) {},
+			expectUpdate: func(r *namespaceservice.UpdateNamespaceRequest) {
+				r.Spec.CodecSpec = &namespace.CodecServerPropertySpec{Endpoint: "fakehost:9999"}
+			},
+		},
+		{
+			args:      []string{"n", "ucs", "-n", ns, "-e", "fakehost:9999", "--pass-access-token"},
+			expectGet: func(g *namespaceservice.GetNamespaceResponse) {},
+			expectUpdate: func(r *namespaceservice.UpdateNamespaceRequest) {
+				r.Spec.CodecSpec = &namespace.CodecServerPropertySpec{
+					Endpoint:        "fakehost:9999",
+					PassAccessToken: true,
+				}
+			},
+		},
+		{
+			args:      []string{"n", "ucs", "-n", ns, "-e", "fakehost:9999", "--pat", "--include-credentials"},
+			expectGet: func(g *namespaceservice.GetNamespaceResponse) {},
+			expectUpdate: func(r *namespaceservice.UpdateNamespaceRequest) {
+				r.Spec.CodecSpec = &namespace.CodecServerPropertySpec{
+					Endpoint:           "fakehost:9999",
+					PassAccessToken:    true,
+					IncludeCredentials: true,
+				}
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		s.Run(strings.Join(tc.args, " "), func() {
+			getResp := namespaceservice.GetNamespaceResponse{
+				Namespace: &namespace.Namespace{
+					Namespace: ns,
+					Spec: &namespace.NamespaceSpec{
+						AcceptedClientCa: "cert1",
+						SearchAttributes: map[string]namespace.SearchAttributeType{
+							"attr1": namespace.SEARCH_ATTRIBUTE_TYPE_BOOL,
+						},
+						RetentionDays: 7,
+						CertificateFilters: []*namespace.CertificateFilterSpec{
+							{
+								CommonName: "test0",
+							},
+						},
+					},
+					State:           namespace.STATE_ACTIVE,
+					ResourceVersion: "ver1",
+				},
+			}
+			if tc.expectGet != nil {
+				tc.expectGet(&getResp)
+				s.mockService.EXPECT().GetNamespace(gomock.Any(), &namespaceservice.GetNamespaceRequest{
+					Namespace: ns,
+				}).Return(&getResp, nil).Times(1)
+			}
+
+			if tc.expectUpdate != nil {
+				spec := *(getResp.Namespace.Spec)
+				req := namespaceservice.UpdateNamespaceRequest{
+					Namespace:       ns,
+					Spec:            &spec,
+					ResourceVersion: getResp.Namespace.ResourceVersion,
+				}
+				tc.expectUpdate(&req)
+				s.mockService.EXPECT().UpdateNamespace(gomock.Any(), &req).
+					Return(&namespaceservice.UpdateNamespaceResponse{
+						RequestStatus: &request.RequestStatus{},
+					}, nil).Times(1)
+			}
+
+			err := s.RunCmd(tc.args...)
+			if tc.expectErr {
+				s.Error(err)
+			} else {
+				s.NoError(err)
+			}
+		})
+	}
+}
+
 func (s *NamespaceTestSuite) TestUpdateNamespaceRetention() {
 
 	ns := "ns1"
@@ -1306,4 +1417,398 @@ func (s *NamespaceTestSuite) TestDelete() {
 		RequestStatus: &request.RequestStatus{},
 	}, nil).Times(1)
 	s.NoError(s.RunCmd("namespace", "delete", "--namespace", "ns1"))
+}
+
+func (s *NamespaceTestSuite) TestCreateExportSink() {
+	ns := "testNamespace"
+	type morphGetResp func(*namespaceservice.GetNamespaceResponse)
+	type morphCreateSinkReq func(*namespaceservice.CreateExportSinkRequest)
+
+	tests := []struct {
+		name          string
+		args          []string
+		expectGet     morphGetResp
+		expectRequest morphCreateSinkReq
+		expectErr     bool
+		expectErrMsg  string
+	}{
+		{
+			name:      "create export sink",
+			args:      []string{"namespace", "es", "create", "--namespace", ns, "--sink-name", "sink1", "--role-arn", "arn:aws:iam::123456789012:role/TestRole", "--s3-bucket-name", "testBucket"},
+			expectGet: func(g *namespaceservice.GetNamespaceResponse) {},
+			expectRequest: func(r *namespaceservice.CreateExportSinkRequest) {
+				r.Namespace = ns
+				r.Spec = &sink.ExportSinkSpec{
+					Name:            "sink1",
+					Enabled:         true,
+					DestinationType: sink.EXPORT_DESTINATION_TYPE_S3,
+					S3Sink: &sink.S3Spec{
+						RoleName:     "TestRole",
+						BucketName:   "testBucket",
+						Region:       "us-west-2",
+						AwsAccountId: "123456789012",
+					},
+				}
+			},
+		},
+		{
+			name:         "create export sink with invalid role arn",
+			args:         []string{"namespace", "es", "create", "--namespace", ns, "--sink-name", "sink1", "--role-arn", "testRole", "--s3-bucket-name", "testBucket"},
+			expectErr:    true,
+			expectErrMsg: "invalid assumed role: testRole",
+		},
+		{
+			name: "create export sink with invalid namespace",
+			args: []string{"namespace", "es", "create", "--namespace", ns, "--sink-name", "sink1", "--role-arn", "arn:aws:iam::123456789012:role/TestRole", "--s3-bucket-name", "testBucket"},
+			expectGet: func(g *namespaceservice.GetNamespaceResponse) {
+				g.Namespace = &namespace.Namespace{
+					Namespace: "",
+					Spec: &namespace.NamespaceSpec{
+						Region: "us-west-2",
+					},
+				}
+			},
+			expectErr:    true,
+			expectErrMsg: "unable to get namespace: invalid namespace returned by server",
+		},
+	}
+
+	for _, tc := range tests {
+		s.Run(strings.Join(tc.args, " "), func() {
+			getResp := namespaceservice.GetNamespaceResponse{
+				Namespace: &namespace.Namespace{
+					Namespace: ns,
+					Spec: &namespace.NamespaceSpec{
+						Region: "us-west-2",
+					},
+				},
+			}
+
+			if tc.expectGet != nil {
+				tc.expectGet(&getResp)
+				s.mockService.EXPECT().GetNamespace(gomock.Any(), &namespaceservice.GetNamespaceRequest{
+					Namespace: ns,
+				}).Return(&getResp, nil).Times(1)
+			}
+
+			if tc.expectRequest != nil {
+				req := namespaceservice.CreateExportSinkRequest{}
+				tc.expectRequest(&req)
+				s.mockService.EXPECT().CreateExportSink(gomock.Any(), &req).
+					Return(&namespaceservice.CreateExportSinkResponse{RequestStatus: &request.RequestStatus{}}, nil).Times(1)
+			}
+
+			err := s.RunCmd(tc.args...)
+			if tc.expectErr {
+				s.Error(err)
+				s.ErrorContains(err, tc.expectErrMsg)
+			} else {
+				s.NoError(err)
+			}
+		})
+	}
+}
+
+func (s *NamespaceTestSuite) TestGetExportSink() {
+	ns := "namespace"
+	type morphGetReq func(*namespaceservice.GetExportSinkRequest)
+
+	tests := []struct {
+		name          string
+		args          []string
+		expectRequest morphGetReq
+		expectErr     bool
+	}{
+		{
+			name: "get export sink succeeds",
+			args: []string{"namespace", "es", "get", "--namespace", ns, "--sink-name", "sink1"},
+			expectRequest: func(r *namespaceservice.GetExportSinkRequest) {
+				r.Namespace = ns
+				r.SinkName = "sink1"
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		s.Run(strings.Join(tc.args, " "), func() {
+			if tc.expectRequest != nil {
+				req := namespaceservice.GetExportSinkRequest{}
+				tc.expectRequest(&req)
+				s.mockService.EXPECT().GetExportSink(gomock.Any(), &req).
+					Return(&namespaceservice.GetExportSinkResponse{Sink: &sink.ExportSink{}}, nil).Times(1)
+			}
+
+			err := s.RunCmd(tc.args...)
+			if tc.expectErr {
+				s.Error(err)
+			} else {
+				s.NoError(err)
+			}
+		})
+	}
+}
+
+func (s *NamespaceTestSuite) TestDeleteExportSink() {
+	ns := "namespace"
+	type morphDeleteReq func(*namespaceservice.DeleteExportSinkRequest)
+	type morphGetSinkResp func(*namespaceservice.GetExportSinkResponse)
+
+	tests := []struct {
+		name                  string
+		args                  []string
+		expectGetSinkResponse morphGetSinkResp
+		expectRequest         morphDeleteReq
+		expectErr             bool
+		expectErrMsg          string
+	}{
+		{
+			name: "delete export sink succeeds without resource version",
+			args: []string{"namespace", "es", "delete", "--namespace", ns, "--sink-name", "sink1"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {
+				r.Sink = &sink.ExportSink{
+					ResourceVersion: "124214124",
+				}
+			},
+			expectRequest: func(r *namespaceservice.DeleteExportSinkRequest) {
+				r.Namespace = ns
+				r.SinkName = "sink1"
+				r.ResourceVersion = "124214124"
+			},
+		},
+		{
+			name: "delete export succeeds sink with resource version",
+			args: []string{"namespace", "es", "delete", "--namespace", ns, "--sink-name", "sink1", "--resource-version", "999999999"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {
+				r.Sink = &sink.ExportSink{}
+			},
+			expectRequest: func(r *namespaceservice.DeleteExportSinkRequest) {
+				r.Namespace = ns
+				r.SinkName = "sink1"
+				r.ResourceVersion = "999999999"
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		s.Run(strings.Join(tc.args, " "), func() {
+			if tc.expectGetSinkResponse != nil {
+				getSinkResp := namespaceservice.GetExportSinkResponse{Sink: &sink.ExportSink{}}
+				tc.expectGetSinkResponse(&getSinkResp)
+				s.mockService.EXPECT().GetExportSink(gomock.Any(), gomock.Any()).Return(&getSinkResp, nil).Times(1)
+			}
+
+			if tc.expectRequest != nil {
+				req := namespaceservice.DeleteExportSinkRequest{}
+				tc.expectRequest(&req)
+				s.mockService.EXPECT().DeleteExportSink(gomock.Any(), &req).
+					Return(&namespaceservice.DeleteExportSinkResponse{RequestStatus: &request.RequestStatus{}}, nil).Times(1)
+			}
+
+			err := s.RunCmd(tc.args...)
+			if tc.expectErr {
+				s.Error(err)
+			} else {
+				s.NoError(err)
+			}
+		})
+	}
+}
+
+func (s *NamespaceTestSuite) TestListExportSinks() {
+	ns := "namespace"
+	type morphGetReq func(*namespaceservice.ListExportSinksRequest)
+
+	tests := []struct {
+		name string
+
+		args          []string
+		expectRequest morphGetReq
+		expectErr     bool
+	}{
+		{
+			name: "list export sinks succeeds",
+			args: []string{"namespace", "es", "list", "--namespace", ns},
+			expectRequest: func(r *namespaceservice.ListExportSinksRequest) {
+				r.Namespace = ns
+				r.PageSize = 100
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		s.Run(strings.Join(tc.args, " "), func() {
+			if tc.expectRequest != nil {
+				req := namespaceservice.ListExportSinksRequest{}
+				tc.expectRequest(&req)
+				s.mockService.EXPECT().ListExportSinks(gomock.Any(), &req).
+					Return(&namespaceservice.ListExportSinksResponse{}, nil).Times(1)
+			}
+
+			err := s.RunCmd(tc.args...)
+			if tc.expectErr {
+				s.Error(err)
+			} else {
+				s.NoError(err)
+			}
+		})
+	}
+}
+
+func (s *NamespaceTestSuite) TestUpdateExportSink() {
+	ns := "namespace"
+	type morphGetReq func(*namespaceservice.UpdateExportSinkRequest)
+	type morphGetSinkResp func(*namespaceservice.GetExportSinkResponse)
+
+	tests := []struct {
+		name                  string
+		args                  []string
+		expectGetSinkResponse morphGetSinkResp
+		expectRequest         morphGetReq
+		expectErr             bool
+		expectErrMsg          string
+	}{
+		{
+			name:                  "update export sink succeeds with no input",
+			args:                  []string{"namespace", "es", "update", "--namespace", ns, "--sink-name", "testSink"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {},
+		},
+		{
+			name:                  "update export sink succeeds with no updates",
+			args:                  []string{"namespace", "es", "update", "--namespace", ns, "--role-arn", "arn:aws:iam::123456789012:role/TestRole", "--s3-bucket-name", "testBucket", "--enabled", "true", "--sink-name", "testSink"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {},
+		},
+		{
+			name:         "update export sink succeeds with no updates",
+			args:         []string{"namespace", "es", "update", "--namespace", ns, "--role-arn", "arn:aws:iam::123456789012:role/TestRole", "--s3-bucket-name", "testBucket", "--enabled", "true"},
+			expectErr:    true,
+			expectErrMsg: "Required flag \"sink-name\" not set",
+		},
+		{
+			name:                  "update export sink succeeds with not valid enabled value",
+			args:                  []string{"namespace", "es", "update", "--namespace", ns, "--role-arn", "arn:aws:iam::123456789012:role/TestRole", "--s3-bucket-name", "testBucket", "--sink-name", "testSink", "--enabled", ""},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {},
+			expectErr:             true,
+			expectErrMsg:          "invalid value for enabled flag",
+		},
+		{
+			name:                  "update export sink succeeds with enable flag",
+			args:                  []string{"namespace", "es", "update", "--namespace", ns, "--enabled", "false", "--sink-name", "testSink"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {},
+			expectRequest: func(r *namespaceservice.UpdateExportSinkRequest) {
+				r.Namespace = ns
+				r.Spec = &sink.ExportSinkSpec{
+					Name:            "sink1",
+					Enabled:         false,
+					DestinationType: sink.EXPORT_DESTINATION_TYPE_S3,
+					S3Sink: &sink.S3Spec{
+						RoleName:     "TestRole",
+						BucketName:   "testBucket",
+						Region:       "us-west-2",
+						AwsAccountId: "123456789012",
+					},
+				}
+				r.ResourceVersion = "124214124"
+			},
+		},
+		{
+			name:                  "update export sink succeeds with role arn and enabled flag",
+			args:                  []string{"namespace", "es", "update", "--namespace", ns, "--enabled", "false", "--role-arn", "arn:aws:iam::923456789012:role/newTestRole", "--sink-name", "testSink"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {},
+			expectRequest: func(r *namespaceservice.UpdateExportSinkRequest) {
+				r.Namespace = ns
+				r.Spec = &sink.ExportSinkSpec{
+					Name:            "sink1",
+					Enabled:         false,
+					DestinationType: sink.EXPORT_DESTINATION_TYPE_S3,
+					S3Sink: &sink.S3Spec{
+						RoleName:     "newTestRole",
+						BucketName:   "testBucket",
+						Region:       "us-west-2",
+						AwsAccountId: "923456789012",
+					},
+				}
+				r.ResourceVersion = "124214124"
+			},
+		},
+		{
+			name:                  "update export sink succeeds with role arn, bucket name and enabled flag",
+			args:                  []string{"namespace", "es", "update", "--namespace", ns, "--role-arn", "arn:aws:iam::923456789012:role/newTestRole", "--s3-bucket-name", "newTestBucket", "--enabled", "false", "--sink-name", "testSink"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {},
+			expectRequest: func(r *namespaceservice.UpdateExportSinkRequest) {
+				r.Namespace = ns
+				r.Spec = &sink.ExportSinkSpec{
+					Name:            "sink1",
+					Enabled:         false,
+					DestinationType: sink.EXPORT_DESTINATION_TYPE_S3,
+					S3Sink: &sink.S3Spec{
+						RoleName:     "newTestRole",
+						BucketName:   "newTestBucket",
+						Region:       "us-west-2",
+						AwsAccountId: "923456789012",
+					},
+				}
+				r.ResourceVersion = "124214124"
+			},
+		},
+		{
+			name:                  "update export sink succeeds with role arn, bucket name, kms arn and enabled flag",
+			args:                  []string{"namespace", "es", "update", "--namespace", ns, "--role-arn", "arn:aws:iam::923456789012:role/newTestRole", "--s3-bucket-name", "newTestBucket", "--kms-arn", "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab", "--enabled", "false", "--sink-name", "testSink"},
+			expectGetSinkResponse: func(r *namespaceservice.GetExportSinkResponse) {},
+			expectRequest: func(r *namespaceservice.UpdateExportSinkRequest) {
+				r.Namespace = ns
+				r.Spec = &sink.ExportSinkSpec{
+					Name:            "sink1",
+					Enabled:         false,
+					DestinationType: sink.EXPORT_DESTINATION_TYPE_S3,
+					S3Sink: &sink.S3Spec{
+						RoleName:     "newTestRole",
+						BucketName:   "newTestBucket",
+						Region:       "us-west-2",
+						AwsAccountId: "923456789012",
+						KmsArn:       "arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab",
+					},
+				}
+				r.ResourceVersion = "124214124"
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		s.Run(strings.Join(tc.args, " "), func() {
+			if tc.expectGetSinkResponse != nil {
+				getSinkResp := namespaceservice.GetExportSinkResponse{Sink: &sink.ExportSink{
+					Name: "sink1",
+					Spec: &sink.ExportSinkSpec{
+						Name:            "sink1",
+						Enabled:         true,
+						DestinationType: sink.EXPORT_DESTINATION_TYPE_S3,
+						S3Sink: &sink.S3Spec{
+							RoleName:     "TestRole",
+							BucketName:   "testBucket",
+							Region:       "us-west-2",
+							AwsAccountId: "123456789012",
+						},
+					},
+					ResourceVersion: "124214124",
+				}}
+				tc.expectGetSinkResponse(&getSinkResp)
+				s.mockService.EXPECT().GetExportSink(gomock.Any(), gomock.Any()).Return(&getSinkResp, nil).Times(1)
+			}
+
+			if tc.expectRequest != nil {
+				req := namespaceservice.UpdateExportSinkRequest{}
+				tc.expectRequest(&req)
+				s.mockService.EXPECT().UpdateExportSink(gomock.Any(), &req).
+					Return(&namespaceservice.UpdateExportSinkResponse{RequestStatus: &request.RequestStatus{}}, nil).Times(1)
+			}
+
+			err := s.RunCmd(tc.args...)
+			if tc.expectErr {
+				s.Error(err)
+				s.ErrorContains(err, tc.expectErrMsg)
+			} else {
+				s.NoError(err)
+			}
+		})
+	}
 }
