@@ -5,15 +5,15 @@ package accountservice
 
 import (
 	fmt "fmt"
+	proto "github.com/gogo/protobuf/proto"
+	types "github.com/gogo/protobuf/types"
 	io "io"
 	math "math"
 	math_bits "math/bits"
 	reflect "reflect"
 	strings "strings"
-
-	proto "github.com/gogo/protobuf/proto"
-	types "github.com/gogo/protobuf/types"
 	v1 "github.com/temporalio/tcld/protogen/api/account/v1"
+	v12 "github.com/temporalio/tcld/protogen/api/common/v1"
 	v11 "github.com/temporalio/tcld/protogen/api/request/v1"
 )
 
@@ -547,6 +547,84 @@ func (m *UpdateAccountResponse) GetRequestStatus() *v11.RequestStatus {
 	return nil
 }
 
+type GetRegionsRequest struct {
+}
+
+func (m *GetRegionsRequest) Reset()      { *m = GetRegionsRequest{} }
+func (*GetRegionsRequest) ProtoMessage() {}
+func (*GetRegionsRequest) Descriptor() ([]byte, []int) {
+	return fileDescriptor_e50566fc74ec64cc, []int{10}
+}
+func (m *GetRegionsRequest) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *GetRegionsRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_GetRegionsRequest.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *GetRegionsRequest) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_GetRegionsRequest.Merge(m, src)
+}
+func (m *GetRegionsRequest) XXX_Size() int {
+	return m.Size()
+}
+func (m *GetRegionsRequest) XXX_DiscardUnknown() {
+	xxx_messageInfo_GetRegionsRequest.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_GetRegionsRequest proto.InternalMessageInfo
+
+type GetRegionsResponse struct {
+	Regions []*v12.Region `protobuf:"bytes,1,rep,name=regions,proto3" json:"regions,omitempty"`
+}
+
+func (m *GetRegionsResponse) Reset()      { *m = GetRegionsResponse{} }
+func (*GetRegionsResponse) ProtoMessage() {}
+func (*GetRegionsResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_e50566fc74ec64cc, []int{11}
+}
+func (m *GetRegionsResponse) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *GetRegionsResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_GetRegionsResponse.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *GetRegionsResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_GetRegionsResponse.Merge(m, src)
+}
+func (m *GetRegionsResponse) XXX_Size() int {
+	return m.Size()
+}
+func (m *GetRegionsResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_GetRegionsResponse.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_GetRegionsResponse proto.InternalMessageInfo
+
+func (m *GetRegionsResponse) GetRegions() []*v12.Region {
+	if m != nil {
+		return m.Regions
+	}
+	return nil
+}
+
 func init() {
 	proto.RegisterType((*GetNamespaceUsageRequest)(nil), "api.accountservice.v1.GetNamespaceUsageRequest")
 	proto.RegisterType((*GetNamespaceUsageResponse)(nil), "api.accountservice.v1.GetNamespaceUsageResponse")
@@ -558,6 +636,8 @@ func init() {
 	proto.RegisterType((*GetAccountResponse)(nil), "api.accountservice.v1.GetAccountResponse")
 	proto.RegisterType((*UpdateAccountRequest)(nil), "api.accountservice.v1.UpdateAccountRequest")
 	proto.RegisterType((*UpdateAccountResponse)(nil), "api.accountservice.v1.UpdateAccountResponse")
+	proto.RegisterType((*GetRegionsRequest)(nil), "api.accountservice.v1.GetRegionsRequest")
+	proto.RegisterType((*GetRegionsResponse)(nil), "api.accountservice.v1.GetRegionsResponse")
 }
 
 func init() {
@@ -565,48 +645,50 @@ func init() {
 }
 
 var fileDescriptor_e50566fc74ec64cc = []byte{
-	// 644 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x95, 0x4d, 0x4f, 0xd4, 0x5e,
-	0x14, 0xc6, 0xe7, 0x32, 0xfc, 0xff, 0xd0, 0x83, 0xbc, 0x55, 0x81, 0x91, 0x97, 0xeb, 0xa4, 0x31,
-	0x66, 0x4c, 0x4c, 0xeb, 0x60, 0x8c, 0x31, 0xae, 0x7c, 0x0b, 0x61, 0xa1, 0x21, 0x05, 0x5c, 0x98,
-	0x98, 0xe6, 0xd2, 0x1e, 0x9a, 0x46, 0xa6, 0xad, 0xf7, 0xde, 0x4e, 0x94, 0x95, 0xdf, 0x40, 0xfd,
-	0x16, 0x7e, 0x14, 0x5d, 0xc9, 0x92, 0xa5, 0x94, 0x8d, 0x3b, 0xf9, 0x08, 0xa6, 0xed, 0x6d, 0x66,
-	0x5a, 0x30, 0x2a, 0x2c, 0xef, 0x73, 0xce, 0xf3, 0x9c, 0x7b, 0x7e, 0xbd, 0x30, 0x70, 0x8b, 0xc5,
-	0x81, 0xc5, 0x5c, 0x37, 0x4a, 0x42, 0x29, 0x90, 0xf7, 0x03, 0x17, 0xad, 0x7e, 0xd7, 0xe2, 0xf8,
-	0x26, 0x41, 0x21, 0x1d, 0x8e, 0x22, 0x8e, 0x42, 0x81, 0x66, 0xcc, 0x23, 0x19, 0xe9, 0x73, 0x2c,
-	0x0e, 0xcc, 0x6a, 0xb7, 0xd9, 0xef, 0x2e, 0x2e, 0x0f, 0x85, 0x64, 0xee, 0x1e, 0x0a, 0xc1, 0x7c,
-	0x65, 0x2a, 0xaa, 0x2a, 0xf0, 0x74, 0xf5, 0x9a, 0x1f, 0x45, 0xfe, 0x1e, 0x5a, 0xf9, 0x69, 0x27,
-	0xd9, 0xb5, 0x64, 0xd0, 0x43, 0x21, 0x59, 0x2f, 0x2e, 0x1a, 0x8c, 0x9f, 0x04, 0x5a, 0x6b, 0x28,
-	0x9f, 0xb3, 0x1e, 0x8a, 0x98, 0xb9, 0xb8, 0x9d, 0x99, 0xed, 0x22, 0x4e, 0x5f, 0x06, 0x2d, 0x2c,
-	0x0b, 0x2d, 0xd2, 0x26, 0x1d, 0xcd, 0x1e, 0x08, 0xfa, 0x7d, 0x00, 0x21, 0x19, 0x97, 0x4e, 0x96,
-	0xd9, 0x1a, 0x69, 0x93, 0xce, 0xc4, 0xea, 0xa2, 0x59, 0x0c, 0x34, 0xcb, 0x81, 0xe6, 0x56, 0x39,
-	0xd0, 0xd6, 0xf2, 0xee, 0xec, 0xac, 0xdf, 0x85, 0x71, 0x0c, 0xbd, 0xc2, 0xd8, 0xfc, 0xa3, 0x71,
-	0x0c, 0x43, 0x2f, 0xb7, 0x3d, 0x86, 0x59, 0xe6, 0xfb, 0x1c, 0x7d, 0x26, 0xd1, 0x73, 0x62, 0xe4,
-	0x41, 0xe4, 0xb5, 0x46, 0xdb, 0xa4, 0x33, 0xb5, 0x3a, 0x6f, 0x0e, 0xc1, 0x33, 0xfb, 0x5d, 0x73,
-	0x23, 0xaf, 0xda, 0x33, 0x03, 0x43, 0xa1, 0x18, 0x5b, 0x70, 0xf5, 0x8c, 0x85, 0x8b, 0x0f, 0xa1,
-	0xdf, 0x83, 0x31, 0x0c, 0x25, 0x0f, 0x50, 0xb4, 0x48, 0xbb, 0xd9, 0x99, 0x58, 0x5d, 0xa9, 0xe7,
-	0x3e, 0x43, 0x89, 0x3c, 0x08, 0xfd, 0xa7, 0xa1, 0xe4, 0xef, 0xec, 0xb2, 0xdb, 0xf8, 0x4a, 0xaa,
-	0xb1, 0xa2, 0x02, 0x72, 0x09, 0xb4, 0x98, 0xf9, 0xe8, 0x88, 0x60, 0xbf, 0x00, 0xf9, 0x9f, 0x3d,
-	0x9e, 0x09, 0x9b, 0xc1, 0x3e, 0xea, 0x2b, 0x00, 0x79, 0x51, 0x46, 0xaf, 0x31, 0xcc, 0x39, 0x6a,
-	0x76, 0xde, 0xbe, 0x95, 0x09, 0x35, 0xcc, 0xcd, 0xf3, 0x62, 0x1e, 0xfd, 0x6b, 0xcc, 0xc6, 0x07,
-	0x02, 0x8b, 0x67, 0xed, 0xa2, 0x18, 0xad, 0xc3, 0x64, 0x92, 0x09, 0x0e, 0x47, 0x37, 0xe2, 0x5e,
-	0x49, 0xea, 0x7a, 0x9d, 0x54, 0x1d, 0x71, 0xd6, 0x6c, 0x5f, 0x4a, 0x06, 0x07, 0xa1, 0xdf, 0x80,
-	0xe9, 0x10, 0xdf, 0x4a, 0xe7, 0xd4, 0xfe, 0x93, 0x99, 0xbc, 0x51, 0x32, 0x30, 0xbe, 0x11, 0x98,
-	0x5f, 0x43, 0xf9, 0xb0, 0x08, 0xaf, 0xa0, 0xad, 0xe2, 0x21, 0xe7, 0xc5, 0x33, 0x72, 0xc1, 0x57,
-	0xd8, 0xfc, 0xc7, 0x57, 0x68, 0xc3, 0xc2, 0xa9, 0x85, 0x2e, 0xfa, 0x06, 0x2f, 0xc3, 0xec, 0x20,
-	0x53, 0xf1, 0x31, 0xd6, 0x40, 0x1f, 0x16, 0xd5, 0x8c, 0x2e, 0x8c, 0xa9, 0x3c, 0x85, 0x6c, 0xa1,
-	0x3e, 0xa3, 0x74, 0x94, 0x7d, 0xc6, 0x27, 0x02, 0x57, 0xb6, 0x63, 0x8f, 0x49, 0xac, 0x4e, 0xd0,
-	0x2d, 0x18, 0x15, 0x31, 0xba, 0x2a, 0x68, 0xe9, 0x37, 0x41, 0x9b, 0x31, 0xba, 0x76, 0xde, 0xa8,
-	0xdf, 0x84, 0x19, 0x8e, 0x22, 0x4a, 0xb8, 0x8b, 0x4e, 0x1f, 0xb9, 0x08, 0xa2, 0xf2, 0xb3, 0x4f,
-	0x97, 0xfa, 0x8b, 0x42, 0xce, 0xfe, 0x36, 0xca, 0x7f, 0x96, 0x41, 0x01, 0x59, 0xb3, 0x35, 0xa5,
-	0xac, 0x7b, 0xc6, 0x2b, 0x98, 0xab, 0x5d, 0x49, 0xed, 0xf7, 0x04, 0xa6, 0x4a, 0x9f, 0x90, 0x4c,
-	0x26, 0x42, 0xdd, 0xae, 0x40, 0xa9, 0x4a, 0xd9, 0xed, 0xd4, 0x12, 0x9b, 0x79, 0x93, 0x3d, 0xc9,
-	0x87, 0x8f, 0x8f, 0x76, 0x0f, 0x8e, 0x68, 0xe3, 0xf0, 0x88, 0x36, 0x4e, 0x8e, 0x28, 0x79, 0x9f,
-	0x52, 0xf2, 0x39, 0xa5, 0xe4, 0x4b, 0x4a, 0xc9, 0x41, 0x4a, 0xc9, 0xf7, 0x94, 0x92, 0x1f, 0x29,
-	0x6d, 0x9c, 0xa4, 0x94, 0x7c, 0x3c, 0xa6, 0x8d, 0x83, 0x63, 0xda, 0x38, 0x3c, 0xa6, 0x8d, 0x97,
-	0xb7, 0x65, 0x2f, 0xe6, 0x7b, 0xa6, 0xbb, 0x17, 0x25, 0x9e, 0x75, 0xe6, 0x0f, 0xc0, 0x83, 0xaa,
-	0xb2, 0xf3, 0x7f, 0xfe, 0xdc, 0xee, 0xfc, 0x0a, 0x00, 0x00, 0xff, 0xff, 0x78, 0x3f, 0x0a, 0x36,
-	0x2f, 0x06, 0x00, 0x00,
+	// 688 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x95, 0xcd, 0x6e, 0xd3, 0x4e,
+	0x14, 0xc5, 0x33, 0x4d, 0xff, 0xff, 0x34, 0x53, 0xfa, 0x65, 0x68, 0x1b, 0xd2, 0xd6, 0x44, 0x16,
+	0x42, 0x41, 0x42, 0x36, 0x29, 0x42, 0x08, 0xb1, 0xe2, 0xa3, 0xaa, 0xba, 0x00, 0x55, 0xd3, 0x96,
+	0x05, 0x12, 0x8a, 0x5c, 0xfb, 0xd6, 0xb2, 0xa8, 0x3d, 0x66, 0x66, 0x1c, 0x41, 0x57, 0xbc, 0x01,
+	0xf0, 0x16, 0x3c, 0x0a, 0xac, 0xe8, 0xb2, 0x4b, 0xea, 0x6e, 0xd8, 0xd1, 0x47, 0x40, 0xf6, 0xcc,
+	0x90, 0x38, 0x29, 0x02, 0xda, 0xe5, 0x9c, 0x7b, 0xef, 0x39, 0xbe, 0x3f, 0x4f, 0x1c, 0x7c, 0xcb,
+	0x4d, 0x42, 0xc7, 0xf5, 0x3c, 0x9a, 0xc6, 0x82, 0x03, 0xeb, 0x85, 0x1e, 0x38, 0xbd, 0x8e, 0xc3,
+	0xe0, 0x75, 0x0a, 0x5c, 0x74, 0x19, 0xf0, 0x84, 0xc6, 0x1c, 0xec, 0x84, 0x51, 0x41, 0x8d, 0x79,
+	0x37, 0x09, 0xed, 0x72, 0xb7, 0xdd, 0xeb, 0x34, 0x97, 0x07, 0x4c, 0xf2, 0xe9, 0x08, 0x38, 0x77,
+	0x03, 0x35, 0xd4, 0x5c, 0xca, 0xab, 0x1e, 0x8d, 0x22, 0x1a, 0x8f, 0x16, 0x8b, 0x51, 0x95, 0x36,
+	0x5a, 0xbd, 0x16, 0x50, 0x1a, 0xec, 0x83, 0x53, 0x9c, 0x76, 0xd3, 0x3d, 0x47, 0x84, 0x11, 0x70,
+	0xe1, 0x46, 0x89, 0x6c, 0xb0, 0x7e, 0x20, 0xdc, 0x58, 0x07, 0xf1, 0xcc, 0x8d, 0x80, 0x27, 0xae,
+	0x07, 0x3b, 0xf9, 0x30, 0x91, 0x76, 0xc6, 0x32, 0xae, 0xc7, 0xba, 0xd0, 0x40, 0x2d, 0xd4, 0xae,
+	0x93, 0xbe, 0x60, 0xdc, 0xc7, 0x98, 0x0b, 0x97, 0x89, 0x6e, 0xee, 0xd9, 0x18, 0x6b, 0xa1, 0xf6,
+	0xe4, 0x6a, 0xd3, 0x96, 0x81, 0xb6, 0x0e, 0xb4, 0xb7, 0x75, 0x20, 0xa9, 0x17, 0xdd, 0xf9, 0xd9,
+	0xb8, 0x8b, 0x27, 0x20, 0xf6, 0xe5, 0x60, 0xf5, 0x8f, 0x83, 0x35, 0x88, 0xfd, 0x62, 0xec, 0x31,
+	0x9e, 0x73, 0x83, 0x80, 0x41, 0xe0, 0x0a, 0xf0, 0xbb, 0x09, 0xb0, 0x90, 0xfa, 0x8d, 0xf1, 0x16,
+	0x6a, 0x4f, 0xaf, 0x2e, 0xd8, 0x03, 0x64, 0xed, 0x5e, 0xc7, 0xde, 0x2c, 0xaa, 0x64, 0xb6, 0x3f,
+	0x20, 0x15, 0x6b, 0x1b, 0x5f, 0x3d, 0x63, 0x61, 0xf9, 0x96, 0x8c, 0x7b, 0xb8, 0x06, 0xb1, 0x60,
+	0x21, 0xf0, 0x06, 0x6a, 0x55, 0xdb, 0x93, 0xab, 0x2b, 0xc3, 0xbe, 0x4f, 0x41, 0x00, 0x0b, 0xe3,
+	0x60, 0x2d, 0x16, 0xec, 0x2d, 0xd1, 0xdd, 0xd6, 0x17, 0x54, 0xb6, 0xe5, 0x25, 0x90, 0x4b, 0xb8,
+	0x9e, 0xb8, 0x01, 0x74, 0x79, 0x78, 0x20, 0x41, 0xfe, 0x47, 0x26, 0x72, 0x61, 0x2b, 0x3c, 0x00,
+	0x63, 0x05, 0xe3, 0xa2, 0x28, 0xe8, 0x2b, 0x88, 0x0b, 0x8e, 0x75, 0x52, 0xb4, 0x6f, 0xe7, 0xc2,
+	0x10, 0xe6, 0xea, 0x79, 0x31, 0x8f, 0xff, 0x35, 0x66, 0xeb, 0x3d, 0xc2, 0xcd, 0xb3, 0x76, 0x51,
+	0x8c, 0x36, 0xf0, 0x54, 0x9a, 0x0b, 0x5d, 0x06, 0x1e, 0x65, 0xbe, 0x26, 0x75, 0x7d, 0x98, 0xd4,
+	0x30, 0xe2, 0xbc, 0x99, 0x5c, 0x4a, 0xfb, 0x07, 0x6e, 0xdc, 0xc0, 0x33, 0x31, 0xbc, 0x11, 0xdd,
+	0x91, 0xfd, 0xa7, 0x72, 0x79, 0x53, 0x33, 0xb0, 0xbe, 0x22, 0xbc, 0xb0, 0x0e, 0xe2, 0xa1, 0x34,
+	0x2f, 0xa1, 0x2d, 0xe3, 0x41, 0xe7, 0xc5, 0x33, 0x76, 0xc1, 0x5b, 0x58, 0xfd, 0xc7, 0x5b, 0x48,
+	0xf0, 0xe2, 0xc8, 0x42, 0x17, 0xbd, 0x83, 0x97, 0xf1, 0x5c, 0xdf, 0x53, 0xf1, 0xb1, 0xd6, 0xb1,
+	0x31, 0x28, 0xaa, 0x8c, 0x0e, 0xae, 0x29, 0x3f, 0x85, 0x6c, 0x71, 0x38, 0x43, 0x4f, 0xe8, 0x3e,
+	0xeb, 0x23, 0xc2, 0x57, 0x76, 0x12, 0xdf, 0x15, 0x50, 0x4e, 0x30, 0x1c, 0x3c, 0xce, 0x13, 0xf0,
+	0x94, 0xd1, 0xd2, 0x6f, 0x8c, 0xb6, 0x12, 0xf0, 0x48, 0xd1, 0x68, 0xdc, 0xc4, 0xb3, 0x0c, 0x38,
+	0x4d, 0x99, 0x07, 0xdd, 0x1e, 0x30, 0x1e, 0x52, 0xfd, 0xda, 0x67, 0xb4, 0xfe, 0x5c, 0xca, 0xf9,
+	0x6f, 0x43, 0x7f, 0x49, 0x43, 0x09, 0xb9, 0x4e, 0xea, 0x4a, 0xd9, 0xf0, 0xad, 0x97, 0x78, 0x7e,
+	0xe8, 0x91, 0xd4, 0x7e, 0x4f, 0xf0, 0xb4, 0x9e, 0xe3, 0xc2, 0x15, 0x29, 0x57, 0x4f, 0x27, 0x51,
+	0xaa, 0x52, 0xfe, 0x74, 0x6a, 0x89, 0xad, 0xa2, 0x89, 0x4c, 0xb1, 0xc1, 0xa3, 0x02, 0x4a, 0x20,
+	0x08, 0x69, 0xcc, 0x35, 0xd0, 0xb5, 0x02, 0xe8, 0x2f, 0x51, 0x05, 0x3a, 0xb8, 0xc6, 0xa4, 0xa4,
+	0x5e, 0xda, 0x7c, 0x91, 0x24, 0xbf, 0xda, 0x32, 0x28, 0xaf, 0x12, 0xdd, 0xf5, 0x68, 0xef, 0xf0,
+	0xd8, 0xac, 0x1c, 0x1d, 0x9b, 0x95, 0xd3, 0x63, 0x13, 0xbd, 0xcb, 0x4c, 0xf4, 0x29, 0x33, 0xd1,
+	0xe7, 0xcc, 0x44, 0x87, 0x99, 0x89, 0xbe, 0x65, 0x26, 0xfa, 0x9e, 0x99, 0x95, 0xd3, 0xcc, 0x44,
+	0x1f, 0x4e, 0xcc, 0xca, 0xe1, 0x89, 0x59, 0x39, 0x3a, 0x31, 0x2b, 0x2f, 0x6e, 0x8b, 0x28, 0x61,
+	0xfb, 0xb6, 0xb7, 0x4f, 0x53, 0xdf, 0x39, 0xf3, 0x9f, 0xe7, 0x41, 0x59, 0xd9, 0xfd, 0xbf, 0xb8,
+	0xca, 0x77, 0x7e, 0x06, 0x00, 0x00, 0xff, 0xff, 0xe3, 0x4b, 0xb6, 0xd6, 0xa8, 0x06, 0x00, 0x00,
 }
 
 func (this *GetNamespaceUsageRequest) Equal(that interface{}) bool {
@@ -894,6 +976,56 @@ func (this *UpdateAccountResponse) Equal(that interface{}) bool {
 	}
 	return true
 }
+func (this *GetRegionsRequest) Equal(that interface{}) bool {
+	if that == nil {
+		return this == nil
+	}
+
+	that1, ok := that.(*GetRegionsRequest)
+	if !ok {
+		that2, ok := that.(GetRegionsRequest)
+		if ok {
+			that1 = &that2
+		} else {
+			return false
+		}
+	}
+	if that1 == nil {
+		return this == nil
+	} else if this == nil {
+		return false
+	}
+	return true
+}
+func (this *GetRegionsResponse) Equal(that interface{}) bool {
+	if that == nil {
+		return this == nil
+	}
+
+	that1, ok := that.(*GetRegionsResponse)
+	if !ok {
+		that2, ok := that.(GetRegionsResponse)
+		if ok {
+			that1 = &that2
+		} else {
+			return false
+		}
+	}
+	if that1 == nil {
+		return this == nil
+	} else if this == nil {
+		return false
+	}
+	if len(this.Regions) != len(that1.Regions) {
+		return false
+	}
+	for i := range this.Regions {
+		if !this.Regions[i].Equal(that1.Regions[i]) {
+			return false
+		}
+	}
+	return true
+}
 func (this *GetNamespaceUsageRequest) GoString() string {
 	if this == nil {
 		return "nil"
@@ -1024,6 +1156,27 @@ func (this *UpdateAccountResponse) GoString() string {
 	s = append(s, "&accountservice.UpdateAccountResponse{")
 	if this.RequestStatus != nil {
 		s = append(s, "RequestStatus: "+fmt.Sprintf("%#v", this.RequestStatus)+",\n")
+	}
+	s = append(s, "}")
+	return strings.Join(s, "")
+}
+func (this *GetRegionsRequest) GoString() string {
+	if this == nil {
+		return "nil"
+	}
+	s := make([]string, 0, 4)
+	s = append(s, "&accountservice.GetRegionsRequest{")
+	s = append(s, "}")
+	return strings.Join(s, "")
+}
+func (this *GetRegionsResponse) GoString() string {
+	if this == nil {
+		return "nil"
+	}
+	s := make([]string, 0, 5)
+	s = append(s, "&accountservice.GetRegionsResponse{")
+	if this.Regions != nil {
+		s = append(s, "Regions: "+fmt.Sprintf("%#v", this.Regions)+",\n")
 	}
 	s = append(s, "}")
 	return strings.Join(s, "")
@@ -1466,6 +1619,66 @@ func (m *UpdateAccountResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	return len(dAtA) - i, nil
 }
 
+func (m *GetRegionsRequest) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *GetRegionsRequest) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *GetRegionsRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	return len(dAtA) - i, nil
+}
+
+func (m *GetRegionsResponse) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *GetRegionsResponse) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *GetRegionsResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if len(m.Regions) > 0 {
+		for iNdEx := len(m.Regions) - 1; iNdEx >= 0; iNdEx-- {
+			{
+				size, err := m.Regions[iNdEx].MarshalToSizedBuffer(dAtA[:i])
+				if err != nil {
+					return 0, err
+				}
+				i -= size
+				i = encodeVarintRequestResponse(dAtA, i, uint64(size))
+			}
+			i--
+			dAtA[i] = 0xa
+		}
+	}
+	return len(dAtA) - i, nil
+}
+
 func encodeVarintRequestResponse(dAtA []byte, offset int, v uint64) int {
 	offset -= sovRequestResponse(v)
 	base := offset
@@ -1650,6 +1863,30 @@ func (m *UpdateAccountResponse) Size() (n int) {
 	return n
 }
 
+func (m *GetRegionsRequest) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	return n
+}
+
+func (m *GetRegionsResponse) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if len(m.Regions) > 0 {
+		for _, e := range m.Regions {
+			l = e.Size()
+			n += 1 + l + sovRequestResponse(uint64(l))
+		}
+	}
+	return n
+}
+
 func sovRequestResponse(x uint64) (n int) {
 	return (math_bits.Len64(x|1) + 6) / 7
 }
@@ -1777,6 +2014,30 @@ func (this *UpdateAccountResponse) String() string {
 	}
 	s := strings.Join([]string{`&UpdateAccountResponse{`,
 		`RequestStatus:` + strings.Replace(fmt.Sprintf("%v", this.RequestStatus), "RequestStatus", "v11.RequestStatus", 1) + `,`,
+		`}`,
+	}, "")
+	return s
+}
+func (this *GetRegionsRequest) String() string {
+	if this == nil {
+		return "nil"
+	}
+	s := strings.Join([]string{`&GetRegionsRequest{`,
+		`}`,
+	}, "")
+	return s
+}
+func (this *GetRegionsResponse) String() string {
+	if this == nil {
+		return "nil"
+	}
+	repeatedStringForRegions := "[]*Region{"
+	for _, f := range this.Regions {
+		repeatedStringForRegions += strings.Replace(fmt.Sprintf("%v", f), "Region", "v12.Region", 1) + ","
+	}
+	repeatedStringForRegions += "}"
+	s := strings.Join([]string{`&GetRegionsResponse{`,
+		`Regions:` + repeatedStringForRegions + `,`,
 		`}`,
 	}, "")
 	return s
@@ -2935,6 +3196,146 @@ func (m *UpdateAccountResponse) Unmarshal(dAtA []byte) error {
 				m.RequestStatus = &v11.RequestStatus{}
 			}
 			if err := m.RequestStatus.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipRequestResponse(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if skippy < 0 {
+				return ErrInvalidLengthRequestResponse
+			}
+			if (iNdEx + skippy) < 0 {
+				return ErrInvalidLengthRequestResponse
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *GetRegionsRequest) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowRequestResponse
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: GetRegionsRequest: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: GetRegionsRequest: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		default:
+			iNdEx = preIndex
+			skippy, err := skipRequestResponse(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if skippy < 0 {
+				return ErrInvalidLengthRequestResponse
+			}
+			if (iNdEx + skippy) < 0 {
+				return ErrInvalidLengthRequestResponse
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *GetRegionsResponse) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowRequestResponse
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: GetRegionsResponse: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: GetRegionsResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Regions", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowRequestResponse
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthRequestResponse
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthRequestResponse
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Regions = append(m.Regions, &v12.Region{})
+			if err := m.Regions[len(m.Regions)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
 			iNdEx = postIndex
