@@ -5,15 +5,14 @@ package auditlog
 
 import (
 	fmt "fmt"
+	proto "github.com/gogo/protobuf/proto"
+	types "github.com/gogo/protobuf/types"
 	io "io"
 	math "math"
 	math_bits "math/bits"
 	reflect "reflect"
 	strconv "strconv"
 	strings "strings"
-
-	proto "github.com/gogo/protobuf/proto"
-	types "github.com/gogo/protobuf/types"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -85,7 +84,7 @@ func (LogLevel) EnumDescriptor() ([]byte, []int) {
 }
 
 type LogRecord struct {
-	// log record id
+	// log record id.
 	Id string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
 	// Version of the log entry, starting from "0".
 	Version string `protobuf:"bytes,2,opt,name=version,proto3" json:"version,omitempty"`
@@ -93,7 +92,7 @@ type LogRecord struct {
 	Category LogCategory `protobuf:"varint,3,opt,name=category,proto3,enum=api.auditlog.v1.LogCategory" json:"category,omitempty"`
 	// when does this log get generated from the source
 	EmitTime *types.Timestamp `protobuf:"bytes,4,opt,name=emit_time,json=emitTime,proto3" json:"emit_time,omitempty"`
-	// customer ip address or server name
+	// customer ip address or server name, if it's "Temporal Internal", the real ip address will be hidden.
 	CallerIpAddress string `protobuf:"bytes,5,opt,name=caller_ip_address,json=callerIpAddress,proto3" json:"caller_ip_address,omitempty"`
 	// which namespace
 	Namespace string `protobuf:"bytes,6,opt,name=namespace,proto3" json:"namespace,omitempty"`
@@ -111,6 +110,12 @@ type LogRecord struct {
 	Region            string               `protobuf:"bytes,12,opt,name=region,proto3" json:"region,omitempty"`
 	ThirdPartyDetails *ThirdPartyLogDetail `protobuf:"bytes,13,opt,name=third_party_details,json=thirdPartyDetails,proto3" json:"third_party_details,omitempty"`
 	Tags              []string             `protobuf:"bytes,14,rep,name=tags,proto3" json:"tags,omitempty"`
+	// If it's customer/user's operator, use email address; if it's temporal's support to run operation on behalf of
+	// the customer, use "Temporal Internal" instead
+	UserEmail string `protobuf:"bytes,15,opt,name=user_email,json=userEmail,proto3" json:"user_email,omitempty"`
+	// Detailed information about the operation.  For example, it could be the invited user email for "InviteUser" operation,
+	// could be "enable audit logging" for UpdateAccount operation
+	OperationDetails string `protobuf:"bytes,16,opt,name=operation_details,json=operationDetails,proto3" json:"operation_details,omitempty"`
 }
 
 func (m *LogRecord) Reset()      { *m = LogRecord{} }
@@ -243,6 +248,20 @@ func (m *LogRecord) GetTags() []string {
 	return nil
 }
 
+func (m *LogRecord) GetUserEmail() string {
+	if m != nil {
+		return m.UserEmail
+	}
+	return ""
+}
+
+func (m *LogRecord) GetOperationDetails() string {
+	if m != nil {
+		return m.OperationDetails
+	}
+	return ""
+}
+
 type ThirdPartyLogDetail struct {
 	Id       string           `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
 	EmitTime *types.Timestamp `protobuf:"bytes,2,opt,name=emit_time,json=emitTime,proto3" json:"emit_time,omitempty"`
@@ -340,49 +359,51 @@ func init() {
 func init() { proto.RegisterFile("api/auditlog/v1/message.proto", fileDescriptor_8e3907a7549dbc61) }
 
 var fileDescriptor_8e3907a7549dbc61 = []byte{
-	// 671 bytes of a gzipped FileDescriptorProto
+	// 701 bytes of a gzipped FileDescriptorProto
 	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x54, 0xc1, 0x4e, 0xdb, 0x4a,
-	0x14, 0xcd, 0x24, 0x01, 0xe2, 0x9b, 0xf7, 0x82, 0x99, 0xe8, 0xf1, 0x06, 0x1e, 0x98, 0x08, 0xb1,
-	0x88, 0x58, 0x38, 0x82, 0xb7, 0x68, 0xa5, 0xae, 0x02, 0x31, 0x28, 0x92, 0x49, 0x90, 0x09, 0xad,
-	0x40, 0x95, 0xac, 0x21, 0x9e, 0xba, 0x23, 0x39, 0xb1, 0x35, 0x9e, 0x04, 0xb1, 0xeb, 0x07, 0x74,
-	0xd1, 0xfe, 0x45, 0x3f, 0xa5, 0x4b, 0x96, 0x6c, 0x2a, 0x95, 0xb0, 0xe9, 0x92, 0x4f, 0xa8, 0x3c,
-	0x4e, 0x02, 0x01, 0x36, 0xdd, 0xf9, 0x9e, 0x73, 0x66, 0xe6, 0xdc, 0x7b, 0xae, 0x0c, 0xeb, 0x34,
-	0xe2, 0x35, 0x3a, 0xf0, 0xb8, 0x0c, 0x42, 0xbf, 0x36, 0xdc, 0xa9, 0xf5, 0x58, 0x1c, 0x53, 0x9f,
-	0x99, 0x91, 0x08, 0x65, 0x88, 0x17, 0x69, 0xc4, 0xcd, 0x09, 0x6d, 0x0e, 0x77, 0x56, 0x37, 0xfc,
-	0x30, 0xf4, 0x03, 0x56, 0x53, 0xf4, 0xc5, 0xe0, 0x43, 0x4d, 0xf2, 0x1e, 0x8b, 0x25, 0xed, 0x45,
-	0xe9, 0x89, 0xcd, 0xaf, 0x79, 0xd0, 0xec, 0xd0, 0x77, 0x58, 0x37, 0x14, 0x1e, 0x2e, 0x41, 0x96,
-	0x7b, 0x04, 0x55, 0x50, 0x55, 0x73, 0xb2, 0xdc, 0xc3, 0x04, 0x16, 0x86, 0x4c, 0xc4, 0x3c, 0xec,
-	0x93, 0xac, 0x02, 0x27, 0x25, 0x7e, 0x0d, 0x85, 0x2e, 0x95, 0xcc, 0x0f, 0xc5, 0x15, 0xc9, 0x55,
-	0x50, 0xb5, 0xb4, 0xbb, 0x66, 0x3e, 0x79, 0xdc, 0xb4, 0x43, 0x7f, 0x7f, 0xac, 0x71, 0xa6, 0x6a,
-	0xfc, 0x0a, 0x34, 0xd6, 0xe3, 0xd2, 0x4d, 0x9c, 0x90, 0x7c, 0x05, 0x55, 0x8b, 0xbb, 0xab, 0x66,
-	0x6a, 0xd3, 0x9c, 0xd8, 0x34, 0x3b, 0x13, 0x9b, 0x4e, 0x21, 0x11, 0x27, 0x25, 0xde, 0x86, 0xa5,
-	0x2e, 0x0d, 0x02, 0x26, 0x5c, 0x1e, 0xb9, 0xd4, 0xf3, 0x04, 0x8b, 0x63, 0x32, 0xa7, 0x6c, 0x2d,
-	0xa6, 0x44, 0x33, 0xaa, 0xa7, 0x30, 0x5e, 0x03, 0xad, 0x4f, 0x7b, 0x2c, 0x8e, 0x68, 0x97, 0x91,
-	0x79, 0xa5, 0x79, 0x00, 0x12, 0x36, 0x8c, 0x98, 0xa0, 0x32, 0x69, 0x6c, 0x21, 0x65, 0xa7, 0x00,
-	0x5e, 0x86, 0xf9, 0x58, 0x52, 0x39, 0x88, 0x49, 0x41, 0x51, 0xe3, 0x0a, 0xd7, 0x60, 0x2e, 0x60,
-	0x43, 0x16, 0x10, 0x4d, 0xf5, 0xbb, 0xf2, 0x52, 0xbf, 0x76, 0x22, 0x70, 0x52, 0x5d, 0x32, 0xbd,
-	0x71, 0x3c, 0x04, 0xd2, 0xe9, 0x8d, 0x4b, 0xbc, 0x0a, 0x05, 0xee, 0xb1, 0xbe, 0xe4, 0xf2, 0x8a,
-	0x14, 0x15, 0x35, 0xad, 0x93, 0xe7, 0x05, 0xf3, 0x13, 0x67, 0x7f, 0xa5, 0xcf, 0xa7, 0x15, 0xee,
-	0x40, 0x59, 0x7e, 0xe4, 0xc2, 0x73, 0x23, 0x2a, 0xe4, 0x95, 0xeb, 0x31, 0x49, 0x79, 0x10, 0x93,
-	0xbf, 0xd5, 0x04, 0xb7, 0x9e, 0x99, 0xe9, 0x24, 0xda, 0xe3, 0x44, 0x6a, 0x87, 0x7e, 0x43, 0x89,
-	0x9d, 0x25, 0x39, 0x05, 0x53, 0x24, 0xc6, 0x18, 0xf2, 0x92, 0xfa, 0x31, 0x29, 0x55, 0x72, 0x55,
-	0xcd, 0x51, 0xdf, 0x9b, 0x3f, 0x10, 0x94, 0x5f, 0x38, 0xfe, 0x6c, 0x3b, 0x66, 0x92, 0xcc, 0xfe,
-	0x41, 0x92, 0x15, 0x28, 0x7a, 0x2c, 0xee, 0x0a, 0x1e, 0xa9, 0x04, 0x72, 0xea, 0xc6, 0xc7, 0x10,
-	0xfe, 0x0f, 0xb4, 0x6e, 0xc0, 0x59, 0x5f, 0xba, 0xdc, 0x53, 0x4b, 0xa2, 0x39, 0x85, 0x14, 0x68,
-	0x7a, 0x78, 0x1d, 0xe0, 0xd9, 0x06, 0x68, 0x7c, 0x9a, 0xfd, 0x06, 0x14, 0x05, 0xbd, 0x74, 0x27,
-	0xa3, 0x4f, 0xd3, 0x07, 0x41, 0x2f, 0x8f, 0x52, 0x64, 0xfb, 0x3d, 0x14, 0x1f, 0xad, 0x26, 0x5e,
-	0x03, 0x62, 0xb7, 0x0f, 0xdd, 0xfd, 0x7a, 0xc7, 0x3a, 0x6c, 0x3b, 0x67, 0xee, 0x69, 0xeb, 0xe4,
-	0xd8, 0xda, 0x6f, 0x1e, 0x34, 0xad, 0x86, 0x9e, 0xc1, 0xcb, 0x80, 0x67, 0xd8, 0x7a, 0xe3, 0xa8,
-	0xd9, 0xd2, 0x11, 0xfe, 0x17, 0xca, 0x33, 0xf8, 0xc9, 0xd9, 0x49, 0xc7, 0x3a, 0xd2, 0xb3, 0xdb,
-	0x9f, 0x11, 0x14, 0x26, 0x9b, 0x80, 0x57, 0xe0, 0x9f, 0x44, 0x65, 0x5b, 0x6f, 0x2d, 0xfb, 0xc9,
-	0xc5, 0x18, 0x4a, 0x0f, 0x54, 0xb3, 0x75, 0xd0, 0xd6, 0x11, 0x2e, 0xc3, 0xe2, 0x03, 0xd6, 0xb0,
-	0xf6, 0x4e, 0x0f, 0xf5, 0xec, 0xac, 0xf0, 0x5d, 0xdd, 0x69, 0xe9, 0xb9, 0x59, 0xa1, 0xe5, 0x38,
-	0x6d, 0x47, 0xcf, 0xcf, 0x82, 0x07, 0xf5, 0x4e, 0xdd, 0xd6, 0xe7, 0xf6, 0xce, 0xaf, 0x6f, 0x8d,
-	0xcc, 0xcd, 0xad, 0x91, 0xb9, 0xbf, 0x35, 0xd0, 0xa7, 0x91, 0x81, 0xbe, 0x8d, 0x0c, 0xf4, 0x7d,
-	0x64, 0xa0, 0xeb, 0x91, 0x81, 0x7e, 0x8e, 0x0c, 0xf4, 0x6b, 0x64, 0x64, 0xee, 0x47, 0x06, 0xfa,
-	0x72, 0x67, 0x64, 0xae, 0xef, 0x8c, 0xcc, 0xcd, 0x9d, 0x91, 0x39, 0xdf, 0x92, 0xbd, 0x48, 0x04,
-	0x66, 0x37, 0x08, 0x07, 0x5e, 0xed, 0xc9, 0x3f, 0xe7, 0xcd, 0xe4, 0xfb, 0x62, 0x5e, 0xa5, 0xfc,
-	0xff, 0xef, 0x00, 0x00, 0x00, 0xff, 0xff, 0x10, 0xdb, 0x01, 0xc9, 0x96, 0x04, 0x00, 0x00,
+	0x14, 0x8d, 0x93, 0x00, 0xf1, 0xcd, 0x7b, 0x89, 0x99, 0xe8, 0xf1, 0x0c, 0x05, 0x13, 0x21, 0x16,
+	0x11, 0x95, 0x1c, 0x41, 0x17, 0xad, 0xd4, 0x55, 0x20, 0x06, 0x45, 0x32, 0x09, 0x32, 0xa1, 0x15,
+	0xa8, 0x92, 0x35, 0xc4, 0x53, 0x77, 0x24, 0x27, 0xb6, 0xc6, 0x93, 0x20, 0x76, 0xfd, 0x80, 0x2e,
+	0xfa, 0x19, 0xfd, 0x94, 0x2e, 0x59, 0xb2, 0x41, 0x2a, 0x61, 0xd3, 0x25, 0x9f, 0x50, 0xcd, 0x38,
+	0x31, 0x04, 0xd8, 0x74, 0x37, 0xf7, 0x9c, 0xe3, 0x99, 0x73, 0xef, 0x3d, 0x32, 0xac, 0xe1, 0x88,
+	0xd6, 0xf1, 0xd0, 0xa3, 0x3c, 0x08, 0xfd, 0xfa, 0x68, 0xbb, 0xde, 0x27, 0x71, 0x8c, 0x7d, 0x62,
+	0x46, 0x2c, 0xe4, 0x21, 0x2a, 0xe3, 0x88, 0x9a, 0x53, 0xda, 0x1c, 0x6d, 0xaf, 0xac, 0xfb, 0x61,
+	0xe8, 0x07, 0xa4, 0x2e, 0xe9, 0xf3, 0xe1, 0xe7, 0x3a, 0xa7, 0x7d, 0x12, 0x73, 0xdc, 0x8f, 0x92,
+	0x2f, 0x36, 0x6e, 0xf2, 0xa0, 0xda, 0xa1, 0xef, 0x90, 0x5e, 0xc8, 0x3c, 0x54, 0x82, 0x2c, 0xf5,
+	0x74, 0xa5, 0xaa, 0xd4, 0x54, 0x27, 0x4b, 0x3d, 0xa4, 0xc3, 0xc2, 0x88, 0xb0, 0x98, 0x86, 0x03,
+	0x3d, 0x2b, 0xc1, 0x69, 0x89, 0xde, 0x41, 0xa1, 0x87, 0x39, 0xf1, 0x43, 0x76, 0xa9, 0xe7, 0xaa,
+	0x4a, 0xad, 0xb4, 0xb3, 0x6a, 0x3e, 0x79, 0xdc, 0xb4, 0x43, 0x7f, 0x6f, 0xa2, 0x71, 0x52, 0x35,
+	0x7a, 0x0b, 0x2a, 0xe9, 0x53, 0xee, 0x0a, 0x27, 0x7a, 0xbe, 0xaa, 0xd4, 0x8a, 0x3b, 0x2b, 0x66,
+	0x62, 0xd3, 0x9c, 0xda, 0x34, 0xbb, 0x53, 0x9b, 0x4e, 0x41, 0x88, 0x45, 0x89, 0xb6, 0x60, 0xb1,
+	0x87, 0x83, 0x80, 0x30, 0x97, 0x46, 0x2e, 0xf6, 0x3c, 0x46, 0xe2, 0x58, 0x9f, 0x93, 0xb6, 0xca,
+	0x09, 0xd1, 0x8a, 0x1a, 0x09, 0x8c, 0x56, 0x41, 0x1d, 0xe0, 0x3e, 0x89, 0x23, 0xdc, 0x23, 0xfa,
+	0xbc, 0xd4, 0x3c, 0x00, 0x82, 0x0d, 0x23, 0xc2, 0x30, 0x17, 0x8d, 0x2d, 0x24, 0x6c, 0x0a, 0xa0,
+	0x25, 0x98, 0x8f, 0x39, 0xe6, 0xc3, 0x58, 0x2f, 0x48, 0x6a, 0x52, 0xa1, 0x3a, 0xcc, 0x05, 0x64,
+	0x44, 0x02, 0x5d, 0x95, 0xfd, 0x2e, 0xbf, 0xd4, 0xaf, 0x2d, 0x04, 0x4e, 0xa2, 0x13, 0xd3, 0x9b,
+	0xac, 0x47, 0x87, 0x64, 0x7a, 0x93, 0x12, 0xad, 0x40, 0x81, 0x7a, 0x64, 0xc0, 0x29, 0xbf, 0xd4,
+	0x8b, 0x92, 0x4a, 0x6b, 0xf1, 0x3c, 0x23, 0xbe, 0x70, 0xf6, 0x4f, 0xf2, 0x7c, 0x52, 0xa1, 0x2e,
+	0x54, 0xf8, 0x17, 0xca, 0x3c, 0x37, 0xc2, 0x8c, 0x5f, 0xba, 0x1e, 0xe1, 0x98, 0x06, 0xb1, 0xfe,
+	0xaf, 0x9c, 0xe0, 0xe6, 0x33, 0x33, 0x5d, 0xa1, 0x3d, 0x12, 0x52, 0x3b, 0xf4, 0x9b, 0x52, 0xec,
+	0x2c, 0xf2, 0x14, 0x4c, 0x90, 0x18, 0x21, 0xc8, 0x73, 0xec, 0xc7, 0x7a, 0xa9, 0x9a, 0xab, 0xa9,
+	0x8e, 0x3c, 0xa3, 0x35, 0x80, 0x61, 0x4c, 0x98, 0x4b, 0xfa, 0x98, 0x06, 0x7a, 0x39, 0x99, 0x8f,
+	0x40, 0x2c, 0x01, 0xa0, 0xd7, 0xb0, 0x98, 0x0e, 0x2b, 0xb5, 0xa1, 0x49, 0x95, 0x96, 0x12, 0x93,
+	0xfb, 0x37, 0x6e, 0x14, 0xa8, 0xbc, 0x60, 0xe5, 0x59, 0xd2, 0x66, 0x52, 0x91, 0xfd, 0x8b, 0x54,
+	0x54, 0xa1, 0xe8, 0x91, 0xb8, 0xc7, 0x68, 0x24, 0xb7, 0x99, 0x93, 0x37, 0x3e, 0x86, 0xd0, 0x2b,
+	0x50, 0x7b, 0x01, 0x25, 0x03, 0xee, 0x52, 0x4f, 0x06, 0x4e, 0x75, 0x0a, 0x09, 0xd0, 0xf2, 0x44,
+	0xaf, 0xcf, 0xd2, 0xa4, 0xd2, 0x34, 0x47, 0xeb, 0x50, 0x64, 0xf8, 0xc2, 0x9d, 0xae, 0x31, 0x49,
+	0x12, 0x30, 0x7c, 0x71, 0x98, 0x20, 0x5b, 0x9f, 0xa0, 0xf8, 0x28, 0xe6, 0x68, 0x15, 0x74, 0xbb,
+	0x73, 0xe0, 0xee, 0x35, 0xba, 0xd6, 0x41, 0xc7, 0x39, 0x75, 0x4f, 0xda, 0xc7, 0x47, 0xd6, 0x5e,
+	0x6b, 0xbf, 0x65, 0x35, 0xb5, 0x0c, 0x5a, 0x02, 0x34, 0xc3, 0x36, 0x9a, 0x87, 0xad, 0xb6, 0xa6,
+	0xa0, 0xff, 0xa1, 0x32, 0x83, 0x1f, 0x9f, 0x1e, 0x77, 0xad, 0x43, 0x2d, 0xbb, 0xf5, 0x4d, 0x81,
+	0xc2, 0x34, 0x55, 0x68, 0x19, 0xfe, 0x13, 0x2a, 0xdb, 0xfa, 0x60, 0xd9, 0x4f, 0x2e, 0x46, 0x50,
+	0x7a, 0xa0, 0x5a, 0xed, 0xfd, 0x8e, 0xa6, 0xa0, 0x0a, 0x94, 0x1f, 0xb0, 0xa6, 0xb5, 0x7b, 0x72,
+	0xa0, 0x65, 0x67, 0x85, 0x1f, 0x1b, 0x4e, 0x5b, 0xcb, 0xcd, 0x0a, 0x2d, 0xc7, 0xe9, 0x38, 0x5a,
+	0x7e, 0x16, 0xdc, 0x6f, 0x74, 0x1b, 0xb6, 0x36, 0xb7, 0x7b, 0x76, 0x75, 0x6b, 0x64, 0xae, 0x6f,
+	0x8d, 0xcc, 0xfd, 0xad, 0xa1, 0x7c, 0x1d, 0x1b, 0xca, 0x8f, 0xb1, 0xa1, 0xfc, 0x1c, 0x1b, 0xca,
+	0xd5, 0xd8, 0x50, 0x7e, 0x8d, 0x0d, 0xe5, 0xf7, 0xd8, 0xc8, 0xdc, 0x8f, 0x0d, 0xe5, 0xfb, 0x9d,
+	0x91, 0xb9, 0xba, 0x33, 0x32, 0xd7, 0x77, 0x46, 0xe6, 0x6c, 0x93, 0xf7, 0x23, 0x16, 0x98, 0xbd,
+	0x20, 0x1c, 0x7a, 0xf5, 0x27, 0xff, 0xaf, 0xf7, 0xd3, 0xf3, 0xf9, 0xbc, 0xdc, 0xf2, 0x9b, 0x3f,
+	0x01, 0x00, 0x00, 0xff, 0xff, 0x40, 0x94, 0xb7, 0xb5, 0xe2, 0x04, 0x00, 0x00,
 }
 
 func (x LogCategory) String() string {
@@ -465,6 +486,12 @@ func (this *LogRecord) Equal(that interface{}) bool {
 			return false
 		}
 	}
+	if this.UserEmail != that1.UserEmail {
+		return false
+	}
+	if this.OperationDetails != that1.OperationDetails {
+		return false
+	}
 	return true
 }
 func (this *ThirdPartyLogDetail) Equal(that interface{}) bool {
@@ -510,7 +537,7 @@ func (this *LogRecord) GoString() string {
 	if this == nil {
 		return "nil"
 	}
-	s := make([]string, 0, 18)
+	s := make([]string, 0, 20)
 	s = append(s, "&auditlog.LogRecord{")
 	s = append(s, "Id: "+fmt.Sprintf("%#v", this.Id)+",\n")
 	s = append(s, "Version: "+fmt.Sprintf("%#v", this.Version)+",\n")
@@ -530,6 +557,8 @@ func (this *LogRecord) GoString() string {
 		s = append(s, "ThirdPartyDetails: "+fmt.Sprintf("%#v", this.ThirdPartyDetails)+",\n")
 	}
 	s = append(s, "Tags: "+fmt.Sprintf("%#v", this.Tags)+",\n")
+	s = append(s, "UserEmail: "+fmt.Sprintf("%#v", this.UserEmail)+",\n")
+	s = append(s, "OperationDetails: "+fmt.Sprintf("%#v", this.OperationDetails)+",\n")
 	s = append(s, "}")
 	return strings.Join(s, "")
 }
@@ -578,6 +607,22 @@ func (m *LogRecord) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
+	if len(m.OperationDetails) > 0 {
+		i -= len(m.OperationDetails)
+		copy(dAtA[i:], m.OperationDetails)
+		i = encodeVarintMessage(dAtA, i, uint64(len(m.OperationDetails)))
+		i--
+		dAtA[i] = 0x1
+		i--
+		dAtA[i] = 0x82
+	}
+	if len(m.UserEmail) > 0 {
+		i -= len(m.UserEmail)
+		copy(dAtA[i:], m.UserEmail)
+		i = encodeVarintMessage(dAtA, i, uint64(len(m.UserEmail)))
+		i--
+		dAtA[i] = 0x7a
+	}
 	if len(m.Tags) > 0 {
 		for iNdEx := len(m.Tags) - 1; iNdEx >= 0; iNdEx-- {
 			i -= len(m.Tags[iNdEx])
@@ -830,6 +875,14 @@ func (m *LogRecord) Size() (n int) {
 			n += 1 + l + sovMessage(uint64(l))
 		}
 	}
+	l = len(m.UserEmail)
+	if l > 0 {
+		n += 1 + l + sovMessage(uint64(l))
+	}
+	l = len(m.OperationDetails)
+	if l > 0 {
+		n += 2 + l + sovMessage(uint64(l))
+	}
 	return n
 }
 
@@ -891,6 +944,8 @@ func (this *LogRecord) String() string {
 		`Region:` + fmt.Sprintf("%v", this.Region) + `,`,
 		`ThirdPartyDetails:` + strings.Replace(this.ThirdPartyDetails.String(), "ThirdPartyLogDetail", "ThirdPartyLogDetail", 1) + `,`,
 		`Tags:` + fmt.Sprintf("%v", this.Tags) + `,`,
+		`UserEmail:` + fmt.Sprintf("%v", this.UserEmail) + `,`,
+		`OperationDetails:` + fmt.Sprintf("%v", this.OperationDetails) + `,`,
 		`}`,
 	}, "")
 	return s
@@ -1376,6 +1431,70 @@ func (m *LogRecord) Unmarshal(dAtA []byte) error {
 				return io.ErrUnexpectedEOF
 			}
 			m.Tags = append(m.Tags, string(dAtA[iNdEx:postIndex]))
+			iNdEx = postIndex
+		case 15:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field UserEmail", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowMessage
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthMessage
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthMessage
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.UserEmail = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 16:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field OperationDetails", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowMessage
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthMessage
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthMessage
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.OperationDetails = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
