@@ -8,6 +8,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	github_com_gogo_protobuf_sortkeys "github.com/gogo/protobuf/sortkeys"
 	types "github.com/gogo/protobuf/types"
+	descriptor "github.com/golang/protobuf/protoc-gen-go/descriptor"
 	io "io"
 	math "math"
 	math_bits "math/bits"
@@ -511,6 +512,8 @@ type Namespace struct {
 	Envelope *NamespaceEnvelope `protobuf:"bytes,9,opt,name=envelope,proto3" json:"envelope,omitempty"`
 	// codec server information saved
 	CodecServerProperty *CodecServerPropertySpec `protobuf:"bytes,10,opt,name=codec_server_property,json=codecServerProperty,proto3" json:"codec_server_property,omitempty"`
+	// allowed principals is a list of principals that allowed to access the private links on the namespace
+	AllowedPrincipals []string `protobuf:"bytes,11,rep,name=allowed_principals,json=allowedPrincipals,proto3" json:"allowed_principals,omitempty"`
 }
 
 func (m *Namespace) Reset()      { *m = Namespace{} }
@@ -608,6 +611,22 @@ func (m *Namespace) GetCodecServerProperty() *CodecServerPropertySpec {
 	return nil
 }
 
+func (m *Namespace) GetAllowedPrincipals() []string {
+	if m != nil {
+		return m.AllowedPrincipals
+	}
+	return nil
+}
+
+var E_ApiVersion = &proto.ExtensionDesc{
+	ExtendedType:  (*descriptor.FileOptions)(nil),
+	ExtensionType: (*string)(nil),
+	Field:         52000,
+	Name:          "api.namespace.v1.api_version",
+	Tag:           "bytes,52000,opt,name=api_version",
+	Filename:      "api/namespace/v1/message.proto",
+}
+
 func init() {
 	proto.RegisterEnum("api.namespace.v1.SearchAttributeType", SearchAttributeType_name, SearchAttributeType_value)
 	proto.RegisterEnum("api.namespace.v1.State", State_name, State_value)
@@ -619,89 +638,96 @@ func init() {
 	proto.RegisterType((*NamespaceURI)(nil), "api.namespace.v1.NamespaceURI")
 	proto.RegisterType((*NamespaceEnvelope)(nil), "api.namespace.v1.NamespaceEnvelope")
 	proto.RegisterType((*Namespace)(nil), "api.namespace.v1.Namespace")
+	proto.RegisterExtension(E_ApiVersion)
 }
 
 func init() { proto.RegisterFile("api/namespace/v1/message.proto", fileDescriptor_56458b48206aa18d) }
 
 var fileDescriptor_56458b48206aa18d = []byte{
-	// 1229 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x56, 0x4d, 0x73, 0x13, 0xc7,
-	0x16, 0xd5, 0x58, 0x92, 0x91, 0xae, 0xc0, 0x1e, 0xb7, 0x30, 0x9e, 0xe7, 0x87, 0x07, 0xa1, 0x57,
-	0x3c, 0x8c, 0xeb, 0x3d, 0x29, 0x98, 0x4a, 0x85, 0x0a, 0x0b, 0x4a, 0x96, 0x86, 0x20, 0x62, 0x24,
-	0xd5, 0x68, 0xe4, 0x40, 0x2a, 0xa9, 0xae, 0x51, 0xab, 0xed, 0x0c, 0x48, 0x33, 0x43, 0x77, 0x4b,
-	0x94, 0xb3, 0xca, 0x22, 0x3f, 0x20, 0x95, 0xaa, 0xfc, 0x87, 0xfc, 0x93, 0x64, 0xc9, 0x92, 0x25,
-	0x88, 0x4d, 0x96, 0x6c, 0xb3, 0x4b, 0x75, 0xcf, 0xe8, 0xc3, 0xb2, 0x20, 0x95, 0x9d, 0xe6, 0x9c,
-	0x73, 0xbb, 0xcf, 0xed, 0xfb, 0x61, 0x83, 0xe9, 0x86, 0x5e, 0xd9, 0x77, 0x07, 0x94, 0x87, 0x2e,
-	0xa1, 0xe5, 0xd1, 0xed, 0xf2, 0x80, 0x72, 0xee, 0x9e, 0xd0, 0x52, 0xc8, 0x02, 0x11, 0x20, 0xdd,
-	0x0d, 0xbd, 0xd2, 0x94, 0x2f, 0x8d, 0x6e, 0x6f, 0x5f, 0x3b, 0x09, 0x82, 0x93, 0x3e, 0x2d, 0x2b,
-	0xbe, 0x3b, 0x3c, 0x2e, 0x0b, 0x6f, 0x40, 0xb9, 0x70, 0x07, 0x61, 0x14, 0x52, 0x7c, 0x93, 0x82,
-	0x4b, 0x8d, 0x49, 0x44, 0x3b, 0xa4, 0x04, 0x5d, 0x81, 0x55, 0x46, 0x4f, 0xbc, 0xc0, 0x37, 0xb4,
-	0x82, 0xb6, 0x9b, 0xb5, 0xe3, 0x2f, 0xf4, 0x3f, 0x40, 0x2e, 0x21, 0x34, 0x14, 0xb4, 0x87, 0x49,
-	0xdf, 0xa3, 0xbe, 0xc0, 0xc4, 0x35, 0x56, 0x94, 0x46, 0x9f, 0x30, 0x55, 0x45, 0x54, 0x5d, 0x74,
-	0x03, 0xd6, 0x18, 0x15, 0xd4, 0x17, 0x5e, 0xe0, 0xe3, 0x9e, 0x7b, 0xca, 0x8d, 0x64, 0x41, 0xdb,
-	0x4d, 0xdb, 0x97, 0xa6, 0x68, 0xcd, 0x3d, 0xe5, 0xa8, 0x0b, 0x1b, 0x9c, 0xba, 0x8c, 0x7c, 0x87,
-	0x5d, 0x21, 0x98, 0xd7, 0x1d, 0x0a, 0xca, 0x8d, 0x54, 0x21, 0xb9, 0x9b, 0xdb, 0xff, 0xb4, 0xb4,
-	0x98, 0x4d, 0xe9, 0x8c, 0xd1, 0x52, 0x5b, 0x05, 0x56, 0xa6, 0x71, 0x96, 0x2f, 0xd8, 0xa9, 0xad,
-	0xf3, 0x05, 0x18, 0x3d, 0x81, 0x3c, 0xa1, 0x4c, 0x78, 0xc7, 0x1e, 0x71, 0x05, 0xc5, 0xc7, 0x5e,
-	0x5f, 0x50, 0xc6, 0x8d, 0xb4, 0xba, 0xe5, 0xe6, 0xf9, 0x5b, 0xaa, 0x33, 0xf1, 0x03, 0xa5, 0x95,
-	0xb7, 0xd9, 0x88, 0x2c, 0xc2, 0x1c, 0xdd, 0x87, 0x1c, 0xf5, 0x47, 0x1e, 0x0b, 0xfc, 0x01, 0xf5,
-	0x85, 0xb1, 0x5a, 0xd0, 0x76, 0xd7, 0xf6, 0x77, 0xce, 0x9f, 0x68, 0xcd, 0x44, 0xf6, 0x7c, 0x04,
-	0x7a, 0x08, 0x40, 0x82, 0x1e, 0x25, 0x98, 0x87, 0x94, 0x18, 0x99, 0x82, 0xb6, 0x9b, 0xdb, 0xbf,
-	0xb5, 0xc4, 0x91, 0xd4, 0xb4, 0x29, 0x1b, 0x51, 0xd6, 0x62, 0x41, 0x48, 0x99, 0x38, 0x55, 0x9e,
-	0xb2, 0x2a, 0x58, 0x55, 0xed, 0x26, 0xac, 0x87, 0x2e, 0xe7, 0xde, 0x88, 0xe2, 0xa8, 0x5e, 0xdc,
-	0xc8, 0x16, 0x92, 0xbb, 0x59, 0x7b, 0x2d, 0x86, 0xed, 0x08, 0xdd, 0x7e, 0x06, 0x9b, 0x4b, 0x1f,
-	0x0e, 0xe9, 0x90, 0x7c, 0x4e, 0x4f, 0xe3, 0xa2, 0xcb, 0x9f, 0xe8, 0x1e, 0xa4, 0x47, 0x6e, 0x7f,
-	0x48, 0x55, 0x91, 0xd7, 0xf6, 0x6f, 0x9c, 0x37, 0xb6, 0x70, 0x92, 0x73, 0x1a, 0x52, 0x3b, 0x8a,
-	0xf9, 0x7c, 0xe5, 0xae, 0xf6, 0x28, 0x95, 0xb9, 0xa0, 0x67, 0x8a, 0x3f, 0x6b, 0xb0, 0xf5, 0x81,
-	0x0c, 0xd0, 0x36, 0x64, 0xa8, 0xdf, 0x0b, 0x03, 0xcf, 0x17, 0xf1, 0xcd, 0xd3, 0x6f, 0xb4, 0x07,
-	0x1b, 0xd2, 0x3b, 0x96, 0xbd, 0xc5, 0x39, 0x16, 0xc1, 0x73, 0xea, 0x2b, 0x2b, 0x19, 0x5b, 0xe5,
-	0x5a, 0x51, 0xb8, 0x23, 0x61, 0x54, 0x86, 0xbc, 0xe7, 0x93, 0xfe, 0xb0, 0x47, 0x31, 0x61, 0xb4,
-	0x27, 0x3b, 0xcc, 0xed, 0x47, 0x3d, 0x97, 0xb1, 0x51, 0x4c, 0x55, 0x67, 0x4c, 0xf1, 0x37, 0x0d,
-	0x36, 0x97, 0x16, 0x1a, 0x5d, 0x83, 0x1c, 0x09, 0x06, 0x83, 0xc0, 0xc7, 0x32, 0xd5, 0xd8, 0x15,
-	0x44, 0x90, 0x6c, 0x40, 0x54, 0x84, 0x8b, 0x01, 0x3b, 0x71, 0x7d, 0xef, 0x7b, 0x57, 0xf6, 0x71,
-	0x3c, 0x02, 0x67, 0x30, 0xe9, 0x67, 0xfe, 0xdb, 0xed, 0xe3, 0xa1, 0xef, 0x09, 0xe5, 0x27, 0x6b,
-	0xa3, 0xb3, 0x54, 0xc7, 0xf7, 0x04, 0xba, 0x0b, 0x06, 0x1f, 0x76, 0x9f, 0x51, 0x22, 0xb0, 0x2b,
-	0xad, 0xf8, 0xae, 0x90, 0xb5, 0x54, 0x16, 0x52, 0x2a, 0xea, 0x4a, 0xcc, 0x57, 0x66, 0xb4, 0xb4,
-	0x53, 0x7c, 0x01, 0x17, 0xa7, 0x73, 0xd1, 0xb1, 0xeb, 0xb2, 0x8e, 0x2f, 0x69, 0x77, 0x52, 0xc7,
-	0x97, 0xb4, 0x8b, 0x10, 0xa4, 0x4e, 0x58, 0x48, 0x62, 0xa3, 0xea, 0x37, 0xba, 0x07, 0xdb, 0xa3,
-	0x90, 0xe0, 0xc9, 0x63, 0x63, 0x4e, 0xd9, 0xc8, 0x23, 0xd1, 0x85, 0xf2, 0xdd, 0x64, 0xeb, 0x6c,
-	0x8d, 0x42, 0x62, 0xc5, 0x82, 0x76, 0xc4, 0xab, 0x9b, 0x8a, 0x87, 0xb0, 0x31, 0xbd, 0xd2, 0xf2,
-	0x47, 0xb4, 0x1f, 0x84, 0x14, 0x7d, 0x06, 0x86, 0x4b, 0x64, 0x46, 0x1c, 0x87, 0x94, 0x61, 0x4e,
-	0x49, 0xe0, 0xf7, 0x70, 0xdf, 0x1b, 0x78, 0x51, 0x69, 0xd3, 0xf6, 0x66, 0xcc, 0xb7, 0x28, 0x6b,
-	0x2b, 0xf6, 0x50, 0x92, 0xc5, 0x3f, 0x93, 0x90, 0x9d, 0x1e, 0x87, 0xae, 0x42, 0x76, 0xda, 0x62,
-	0x71, 0x12, 0x33, 0x00, 0xdd, 0x02, 0x9d, 0x51, 0x1e, 0x0c, 0x19, 0xa1, 0x78, 0x44, 0x19, 0x9f,
-	0xbd, 0xff, 0xfa, 0x04, 0x3f, 0x8a, 0x60, 0x74, 0x07, 0x52, 0x6a, 0xaa, 0x92, 0x6a, 0xaa, 0xae,
-	0xfd, 0xcd, 0x36, 0xb1, 0x95, 0x18, 0xfd, 0x1f, 0xd2, 0x5c, 0xb8, 0x22, 0x7a, 0xf3, 0xb5, 0xfd,
-	0xad, 0x25, 0x2d, 0x2f, 0x69, 0x3b, 0x52, 0xa1, 0x1d, 0x00, 0x46, 0x5f, 0x0c, 0x29, 0x17, 0xd8,
-	0xeb, 0x19, 0xe9, 0xc8, 0x6d, 0x8c, 0xd4, 0x7b, 0xe8, 0x21, 0xa0, 0xbe, 0xcb, 0x05, 0x1e, 0x04,
-	0x3d, 0xef, 0xd8, 0xa3, 0x3d, 0x2c, 0xb7, 0xaf, 0x5a, 0x13, 0xb9, 0xfd, 0xed, 0x52, 0xb4, 0x9a,
-	0x4b, 0x93, 0xd5, 0x5c, 0x72, 0x26, 0xab, 0xd9, 0xd6, 0x65, 0xd4, 0xe3, 0x38, 0x48, 0xc2, 0xe8,
-	0x13, 0x48, 0x0e, 0x99, 0x67, 0x5c, 0x50, 0xa1, 0xe6, 0x47, 0x72, 0xe9, 0xd8, 0x75, 0x5b, 0x4a,
-	0xd1, 0x7d, 0x39, 0x59, 0x51, 0x69, 0x8c, 0xac, 0x0a, 0xfb, 0xcf, 0x47, 0xc2, 0x26, 0x55, 0xb4,
-	0xa7, 0x41, 0xe8, 0x5b, 0xd8, 0x8c, 0x77, 0x93, 0x1a, 0x5b, 0x1c, 0xc6, 0x73, 0x6b, 0xc0, 0x3f,
-	0x5d, 0x53, 0x79, 0x72, 0x9e, 0x78, 0x94, 0xca, 0x64, 0xf4, 0xec, 0xde, 0x8f, 0x2b, 0x90, 0x5f,
-	0xb2, 0x44, 0xd0, 0x0d, 0xb8, 0xde, 0xb6, 0x2a, 0x76, 0xf5, 0x21, 0xae, 0x38, 0x8e, 0x5d, 0x3f,
-	0xe8, 0x38, 0x16, 0x76, 0x9e, 0xb6, 0x2c, 0xdc, 0x69, 0xb4, 0x5b, 0x56, 0xb5, 0xfe, 0xa0, 0x6e,
-	0xd5, 0xf4, 0x04, 0xba, 0x0e, 0x3b, 0xcb, 0x65, 0x5f, 0x5a, 0x4f, 0xbf, 0x6a, 0xda, 0x35, 0x5d,
-	0x43, 0x26, 0x6c, 0x2f, 0x97, 0x38, 0xd6, 0x13, 0x47, 0x5f, 0x41, 0x3b, 0xf0, 0xaf, 0xe5, 0x7c,
-	0xbd, 0xe1, 0xe8, 0x49, 0x54, 0x80, 0xab, 0xcb, 0xe9, 0x5a, 0xb3, 0x73, 0x70, 0x68, 0xe9, 0x29,
-	0x54, 0x04, 0xf3, 0x03, 0x8a, 0x8a, 0x63, 0x39, 0xf5, 0xc7, 0x96, 0x9e, 0xfe, 0xb0, 0x89, 0x83,
-	0x66, 0xf3, 0x50, 0x5f, 0xdd, 0xfb, 0x65, 0x05, 0xd2, 0xaa, 0xb1, 0xd0, 0x26, 0x6c, 0xb4, 0x9d,
-	0x8a, 0xb3, 0x98, 0xe8, 0x65, 0xd0, 0x23, 0xb8, 0x52, 0x75, 0xea, 0x47, 0x15, 0xa7, 0xde, 0xf8,
-	0x42, 0xd7, 0xd0, 0xbf, 0x61, 0xeb, 0x2c, 0xda, 0x6c, 0xe0, 0x07, 0x95, 0xfa, 0xa1, 0x55, 0xd3,
-	0x57, 0x90, 0x0e, 0x17, 0xe7, 0x48, 0x4b, 0x4f, 0x22, 0x04, 0x6b, 0xf1, 0xd9, 0xad, 0x5a, 0x74,
-	0x44, 0x0a, 0x6d, 0x41, 0x7e, 0x0e, 0xb3, 0x26, 0xe1, 0xe9, 0x99, 0xb8, 0x66, 0x1d, 0x5a, 0x4a,
-	0xbc, 0x3a, 0x13, 0x2b, 0x6c, 0x2a, 0xbe, 0x80, 0x36, 0xe0, 0xd2, 0x3c, 0x51, 0xd3, 0x33, 0x33,
-	0xc7, 0xed, 0x4e, 0xbb, 0x65, 0x35, 0x6a, 0xf2, 0x84, 0x2c, 0x32, 0xe0, 0xf2, 0x19, 0x74, 0x72,
-	0x04, 0xa0, 0x3c, 0xac, 0x9f, 0x61, 0xac, 0x9a, 0x9e, 0xdb, 0x7b, 0x0e, 0xb9, 0xb9, 0xbf, 0x9d,
-	0x32, 0x5f, 0xab, 0x71, 0x54, 0xb7, 0x9b, 0x8d, 0xc7, 0x56, 0xc3, 0x59, 0x78, 0xa2, 0x3c, 0xac,
-	0xcf, 0x93, 0x35, 0xeb, 0x48, 0xd7, 0xa4, 0x8b, 0x79, 0xd0, 0xb1, 0xda, 0xb2, 0xe6, 0x0b, 0x68,
-	0xcb, 0x6e, 0xd6, 0xf4, 0xe4, 0xc1, 0x37, 0xaf, 0xde, 0x9a, 0x89, 0xd7, 0x6f, 0xcd, 0xc4, 0xfb,
-	0xb7, 0xa6, 0xf6, 0xc3, 0xd8, 0xd4, 0x7e, 0x1d, 0x9b, 0xda, 0xef, 0x63, 0x53, 0x7b, 0x35, 0x36,
-	0xb5, 0x37, 0x63, 0x53, 0xfb, 0x63, 0x6c, 0x26, 0xde, 0x8f, 0x4d, 0xed, 0xa7, 0x77, 0x66, 0xe2,
-	0xd5, 0x3b, 0x33, 0xf1, 0xfa, 0x9d, 0x99, 0xf8, 0xfa, 0xbf, 0x62, 0x10, 0xb2, 0x7e, 0x89, 0xf4,
-	0x83, 0x61, 0xaf, 0xbc, 0xf8, 0xff, 0xd9, 0xbd, 0xe9, 0x47, 0x77, 0x55, 0xcd, 0xf9, 0x9d, 0xbf,
-	0x02, 0x00, 0x00, 0xff, 0xff, 0x2a, 0xc7, 0xbd, 0x7a, 0xc4, 0x09, 0x00, 0x00,
+	// 1314 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x56, 0x4d, 0x73, 0xd3, 0x46,
+	0x18, 0x8e, 0x62, 0x3b, 0xd8, 0xaf, 0x49, 0xa2, 0x6c, 0x08, 0x11, 0x29, 0x11, 0xc6, 0x1d, 0x4a,
+	0xc8, 0x34, 0x4e, 0x30, 0xd3, 0xc2, 0xc0, 0x81, 0x71, 0x62, 0xa5, 0x98, 0x06, 0xdb, 0x23, 0x2b,
+	0x29, 0x74, 0xa6, 0xa3, 0x51, 0x56, 0x9b, 0x74, 0x41, 0x96, 0xc4, 0x6a, 0x6d, 0x26, 0x3d, 0x75,
+	0x3a, 0xfd, 0x01, 0x9d, 0xce, 0xb4, 0xe7, 0x1e, 0x39, 0xf7, 0x4f, 0xb4, 0x47, 0x8e, 0x1c, 0xc1,
+	0x5c, 0x7a, 0xe4, 0x27, 0x74, 0x76, 0x25, 0x7f, 0xc4, 0x31, 0x74, 0x7a, 0xb3, 0x9e, 0xe7, 0x79,
+	0x77, 0x9f, 0x77, 0xdf, 0x8f, 0x04, 0x74, 0x27, 0xa4, 0x9b, 0xbe, 0xd3, 0x26, 0x51, 0xe8, 0x60,
+	0xb2, 0xd9, 0xbd, 0xb9, 0xd9, 0x26, 0x51, 0xe4, 0x1c, 0x93, 0x52, 0xc8, 0x02, 0x1e, 0x20, 0xd5,
+	0x09, 0x69, 0x69, 0xc0, 0x97, 0xba, 0x37, 0x57, 0xae, 0x1c, 0x07, 0xc1, 0xb1, 0x47, 0x36, 0x25,
+	0x7f, 0xd8, 0x39, 0xda, 0xe4, 0xb4, 0x4d, 0x22, 0xee, 0xb4, 0xc3, 0x38, 0x64, 0xa5, 0x30, 0x2e,
+	0x70, 0x49, 0x84, 0x19, 0x0d, 0x79, 0xc0, 0x62, 0x45, 0xf1, 0x4d, 0x1a, 0x66, 0xeb, 0xfd, 0x33,
+	0x5b, 0x21, 0xc1, 0xe8, 0x22, 0xcc, 0x30, 0x72, 0x4c, 0x03, 0x5f, 0x53, 0x0a, 0xca, 0x5a, 0xce,
+	0x4c, 0xbe, 0xd0, 0xe7, 0x80, 0x1c, 0x8c, 0x49, 0xc8, 0x89, 0x6b, 0x63, 0x8f, 0x12, 0x9f, 0xdb,
+	0xd8, 0xd1, 0xa6, 0xa5, 0x46, 0xed, 0x33, 0x3b, 0x92, 0xd8, 0x71, 0xd0, 0x35, 0x98, 0x63, 0x84,
+	0x13, 0x9f, 0xd3, 0xc0, 0xb7, 0x5d, 0xe7, 0x24, 0xd2, 0x52, 0x05, 0x65, 0x2d, 0x63, 0xce, 0x0e,
+	0xd0, 0xaa, 0x73, 0x12, 0xa1, 0x43, 0x58, 0x88, 0x88, 0xc3, 0xf0, 0xf7, 0xb6, 0xc3, 0x39, 0xa3,
+	0x87, 0x1d, 0x4e, 0x22, 0x2d, 0x5d, 0x48, 0xad, 0xe5, 0xcb, 0x5f, 0x94, 0xc6, 0xf3, 0x2d, 0x9d,
+	0x32, 0x5a, 0x6a, 0xc9, 0xc0, 0xca, 0x20, 0xce, 0xf0, 0x39, 0x3b, 0x31, 0xd5, 0x68, 0x0c, 0x46,
+	0x8f, 0x61, 0x11, 0x13, 0xc6, 0xe9, 0x11, 0xc5, 0x0e, 0x27, 0xf6, 0x11, 0xf5, 0x38, 0x61, 0x91,
+	0x96, 0x91, 0xb7, 0x5c, 0x3f, 0x7b, 0xcb, 0xce, 0x50, 0xbc, 0x2b, 0xb5, 0xe2, 0x36, 0x13, 0xe1,
+	0x71, 0x38, 0x42, 0xf7, 0x21, 0x4f, 0xfc, 0x2e, 0x65, 0x81, 0xdf, 0x26, 0x3e, 0xd7, 0x66, 0x0a,
+	0xca, 0xda, 0x5c, 0x79, 0xf5, 0xec, 0x89, 0xc6, 0x50, 0x64, 0x8e, 0x46, 0xa0, 0x07, 0x00, 0x38,
+	0x70, 0x09, 0xb6, 0xa3, 0x90, 0x60, 0x2d, 0x5b, 0x50, 0xd6, 0xf2, 0xe5, 0x1b, 0x13, 0x1c, 0x09,
+	0x4d, 0x8b, 0xb0, 0x2e, 0x61, 0x4d, 0x16, 0x84, 0x84, 0xf1, 0x13, 0xe9, 0x29, 0x27, 0x83, 0x65,
+	0xd5, 0xae, 0xc3, 0x7c, 0xe8, 0x44, 0x11, 0xed, 0x12, 0x3b, 0xae, 0x57, 0xa4, 0xe5, 0x0a, 0xa9,
+	0xb5, 0x9c, 0x39, 0x97, 0xc0, 0x66, 0x8c, 0xae, 0x3c, 0x85, 0xa5, 0x89, 0x0f, 0x87, 0x54, 0x48,
+	0x3d, 0x23, 0x27, 0x49, 0xd1, 0xc5, 0x4f, 0x74, 0x0f, 0x32, 0x5d, 0xc7, 0xeb, 0x10, 0x59, 0xe4,
+	0xb9, 0xf2, 0xb5, 0xb3, 0xc6, 0xc6, 0x4e, 0xb2, 0x4e, 0x42, 0x62, 0xc6, 0x31, 0x77, 0xa7, 0xef,
+	0x28, 0x0f, 0xd3, 0xd9, 0x73, 0x6a, 0xb6, 0xf8, 0xab, 0x02, 0xcb, 0x1f, 0xc8, 0x00, 0xad, 0x40,
+	0x96, 0xf8, 0x6e, 0x18, 0x50, 0x9f, 0x27, 0x37, 0x0f, 0xbe, 0xd1, 0x3a, 0x2c, 0x08, 0xef, 0xb6,
+	0xe8, 0xad, 0x28, 0xb2, 0x79, 0xf0, 0x8c, 0xf8, 0xd2, 0x4a, 0xd6, 0x94, 0xb9, 0x56, 0x24, 0x6e,
+	0x09, 0x18, 0x6d, 0xc2, 0x22, 0xf5, 0xb1, 0xd7, 0x71, 0x89, 0x8d, 0x19, 0x71, 0x45, 0x87, 0x39,
+	0x5e, 0xdc, 0x73, 0x59, 0x13, 0x25, 0xd4, 0xce, 0x90, 0x29, 0xfe, 0xa5, 0xc0, 0xd2, 0xc4, 0x42,
+	0xa3, 0x2b, 0x90, 0xc7, 0x41, 0xbb, 0x1d, 0xf8, 0xb6, 0x48, 0x35, 0x71, 0x05, 0x31, 0x24, 0x1a,
+	0x10, 0x15, 0xe1, 0x7c, 0xc0, 0x8e, 0x1d, 0x9f, 0xfe, 0xe0, 0x88, 0x3e, 0x4e, 0x46, 0xe0, 0x14,
+	0x26, 0xfc, 0x8c, 0x7e, 0x3b, 0x9e, 0xdd, 0xf1, 0x29, 0x97, 0x7e, 0x72, 0x26, 0x3a, 0x4d, 0xed,
+	0xfb, 0x94, 0xa3, 0x3b, 0xa0, 0x45, 0x9d, 0xc3, 0xa7, 0x04, 0x73, 0xdb, 0x11, 0x56, 0x7c, 0x87,
+	0x8b, 0x5a, 0x4a, 0x0b, 0x69, 0x19, 0x75, 0x31, 0xe1, 0x2b, 0x43, 0x5a, 0xd8, 0x29, 0x3e, 0x87,
+	0xf3, 0x83, 0xb9, 0xd8, 0x37, 0x6b, 0xa2, 0x8e, 0x2f, 0xc8, 0x61, 0xbf, 0x8e, 0x2f, 0xc8, 0x21,
+	0x42, 0x90, 0x3e, 0x66, 0x21, 0x4e, 0x8c, 0xca, 0xdf, 0xe8, 0x1e, 0xac, 0x74, 0x43, 0x6c, 0xf7,
+	0x1f, 0xdb, 0x8e, 0x08, 0xeb, 0x52, 0x1c, 0x5f, 0x28, 0xde, 0x4d, 0xb4, 0xce, 0x72, 0x37, 0xc4,
+	0x46, 0x22, 0x68, 0xc5, 0xbc, 0xbc, 0xa9, 0xb8, 0x07, 0x0b, 0x83, 0x2b, 0x0d, 0xbf, 0x4b, 0xbc,
+	0x20, 0x24, 0xe8, 0x36, 0x68, 0x0e, 0x16, 0x19, 0x45, 0x76, 0x48, 0x98, 0x1d, 0x11, 0x1c, 0xf8,
+	0xae, 0xed, 0xd1, 0x36, 0x8d, 0x4b, 0x9b, 0x31, 0x97, 0x12, 0xbe, 0x49, 0x58, 0x4b, 0xb2, 0x7b,
+	0x82, 0x2c, 0xbe, 0x4c, 0x43, 0x6e, 0x70, 0x1c, 0xba, 0x0c, 0xb9, 0x41, 0x8b, 0x25, 0x49, 0x0c,
+	0x01, 0x74, 0x03, 0x54, 0x46, 0xa2, 0xa0, 0xc3, 0x30, 0xb1, 0xbb, 0x84, 0x45, 0xc3, 0xf7, 0x9f,
+	0xef, 0xe3, 0x07, 0x31, 0x8c, 0x6e, 0x41, 0x5a, 0x4e, 0x55, 0x4a, 0x4e, 0xd5, 0x95, 0xff, 0xd8,
+	0x26, 0xa6, 0x14, 0xa3, 0x0d, 0xc8, 0x44, 0xdc, 0xe1, 0xf1, 0x9b, 0xcf, 0x95, 0x97, 0x27, 0xb4,
+	0xbc, 0xa0, 0xcd, 0x58, 0x85, 0x56, 0x01, 0x18, 0x79, 0xde, 0x21, 0x11, 0xb7, 0xa9, 0xab, 0x65,
+	0x62, 0xb7, 0x09, 0x52, 0x73, 0xd1, 0x03, 0x40, 0x9e, 0x13, 0x71, 0xbb, 0x1d, 0xb8, 0xf4, 0x88,
+	0x12, 0xd7, 0x16, 0xfb, 0x59, 0xae, 0x89, 0x7c, 0x79, 0xa5, 0x14, 0xef, 0xe6, 0x52, 0x7f, 0x37,
+	0x97, 0xac, 0xfe, 0xf2, 0x36, 0x55, 0x11, 0xf5, 0x28, 0x09, 0x12, 0x30, 0xda, 0x82, 0x54, 0x87,
+	0x51, 0xed, 0x9c, 0x0c, 0xd5, 0x3f, 0x92, 0xcb, 0xbe, 0x59, 0x33, 0x85, 0x14, 0xdd, 0x17, 0x93,
+	0x15, 0x97, 0x46, 0xcb, 0xc9, 0xb0, 0x4f, 0x3f, 0x12, 0xd6, 0xaf, 0xa2, 0x39, 0x08, 0x42, 0xdf,
+	0xc1, 0x52, 0xb2, 0x9b, 0xe4, 0xd8, 0xda, 0x61, 0x32, 0xb7, 0x1a, 0xfc, 0xdf, 0x35, 0xb5, 0x88,
+	0xcf, 0x12, 0x68, 0x03, 0x90, 0xe3, 0x79, 0xc1, 0x0b, 0xe2, 0xda, 0x21, 0xa3, 0x3e, 0xa6, 0xa1,
+	0x18, 0xd8, 0xbc, 0x6c, 0xbc, 0x85, 0x84, 0x69, 0x0e, 0x88, 0x87, 0xe9, 0x6c, 0x56, 0xcd, 0xad,
+	0xff, 0x3c, 0x0d, 0x8b, 0x13, 0x76, 0x0e, 0xba, 0x06, 0x57, 0x5b, 0x46, 0xc5, 0xdc, 0x79, 0x60,
+	0x57, 0x2c, 0xcb, 0xac, 0x6d, 0xef, 0x5b, 0x86, 0x6d, 0x3d, 0x69, 0x1a, 0xf6, 0x7e, 0xbd, 0xd5,
+	0x34, 0x76, 0x6a, 0xbb, 0x35, 0xa3, 0xaa, 0x4e, 0xa1, 0xab, 0xb0, 0x3a, 0x59, 0xf6, 0xb5, 0xf1,
+	0xe4, 0x9b, 0x86, 0x59, 0x55, 0x15, 0xa4, 0xc3, 0xca, 0x64, 0x89, 0x65, 0x3c, 0xb6, 0xd4, 0x69,
+	0xb4, 0x0a, 0x97, 0x26, 0xf3, 0xb5, 0xba, 0xa5, 0xa6, 0x50, 0x01, 0x2e, 0x4f, 0xa6, 0xab, 0x8d,
+	0xfd, 0xed, 0x3d, 0x43, 0x4d, 0xa3, 0x22, 0xe8, 0x1f, 0x50, 0x54, 0x2c, 0xc3, 0xaa, 0x3d, 0x32,
+	0xd4, 0xcc, 0x87, 0x4d, 0x6c, 0x37, 0x1a, 0x7b, 0xea, 0xcc, 0xfa, 0x6f, 0xd3, 0x90, 0x91, 0x7d,
+	0x88, 0x96, 0x60, 0xa1, 0x65, 0x55, 0xac, 0xf1, 0x44, 0x2f, 0x80, 0x1a, 0xc3, 0x95, 0x1d, 0xab,
+	0x76, 0x50, 0xb1, 0x6a, 0xf5, 0xaf, 0x54, 0x05, 0x7d, 0x02, 0xcb, 0xa7, 0xd1, 0x46, 0xdd, 0xde,
+	0xad, 0xd4, 0xf6, 0x8c, 0xaa, 0x3a, 0x8d, 0x54, 0x38, 0x3f, 0x42, 0x1a, 0x6a, 0x0a, 0x21, 0x98,
+	0x4b, 0xce, 0x6e, 0x56, 0xe3, 0x23, 0xd2, 0x68, 0x19, 0x16, 0x47, 0x30, 0xa3, 0x1f, 0x9e, 0x19,
+	0x8a, 0xab, 0xc6, 0x9e, 0x21, 0xc5, 0x33, 0x43, 0xb1, 0xc4, 0x06, 0xe2, 0x73, 0x68, 0x01, 0x66,
+	0x47, 0x89, 0xaa, 0x9a, 0x1d, 0x3a, 0x6e, 0xed, 0xb7, 0x9a, 0x46, 0xbd, 0x2a, 0x4e, 0xc8, 0x21,
+	0x0d, 0x2e, 0x9c, 0x42, 0xfb, 0x47, 0x00, 0x5a, 0x84, 0xf9, 0x53, 0x8c, 0x51, 0x55, 0xf3, 0xeb,
+	0xcf, 0x20, 0x3f, 0xf2, 0xa7, 0x56, 0xe4, 0x6b, 0xd4, 0x0f, 0x6a, 0x66, 0xa3, 0xfe, 0xc8, 0xa8,
+	0x5b, 0x63, 0x4f, 0xb4, 0x08, 0xf3, 0xa3, 0x64, 0xd5, 0x38, 0x50, 0x15, 0xe1, 0x62, 0x14, 0xb4,
+	0x8c, 0x96, 0xa8, 0xf9, 0x18, 0xda, 0x34, 0x1b, 0x55, 0x35, 0x75, 0xf7, 0x3e, 0xe4, 0x9d, 0x90,
+	0xf6, 0xb7, 0x10, 0xba, 0x7c, 0x66, 0x9e, 0x77, 0xa9, 0x47, 0x1a, 0xa1, 0xdc, 0x78, 0xda, 0x1f,
+	0xbf, 0xc7, 0xab, 0x1f, 0x9c, 0x90, 0x26, 0x0b, 0x6a, 0xbb, 0xfd, 0xd3, 0x9f, 0x97, 0x66, 0xcb,
+	0x5b, 0xe5, 0x5b, 0x1b, 0x5b, 0xb7, 0x37, 0xca, 0x5f, 0x6e, 0x6c, 0xdd, 0x7c, 0xf5, 0x56, 0x9f,
+	0x7a, 0xfd, 0x56, 0x9f, 0x7a, 0xff, 0x56, 0x57, 0x7e, 0xec, 0xe9, 0xca, 0xcb, 0x9e, 0xae, 0xfc,
+	0xdd, 0xd3, 0x95, 0x57, 0x3d, 0x5d, 0x79, 0xd3, 0xd3, 0x95, 0x7f, 0x7a, 0xfa, 0xd4, 0xfb, 0x9e,
+	0xae, 0xfc, 0xf2, 0x4e, 0x9f, 0x7a, 0xf5, 0x4e, 0x9f, 0x7a, 0xfd, 0x4e, 0x9f, 0xfa, 0xf6, 0x33,
+	0xde, 0x0e, 0x99, 0x57, 0xc2, 0x5e, 0xd0, 0x71, 0x37, 0xc7, 0xff, 0x85, 0xbc, 0x37, 0xf8, 0x38,
+	0x9c, 0x91, 0xc6, 0x6e, 0xfd, 0x1b, 0x00, 0x00, 0xff, 0xff, 0x3f, 0x29, 0x7e, 0x8b, 0x67, 0x0a,
+	0x00, 0x00,
 }
 
 func (x SearchAttributeType) String() string {
@@ -953,6 +979,14 @@ func (this *Namespace) Equal(that interface{}) bool {
 	if !this.CodecServerProperty.Equal(that1.CodecServerProperty) {
 		return false
 	}
+	if len(this.AllowedPrincipals) != len(that1.AllowedPrincipals) {
+		return false
+	}
+	for i := range this.AllowedPrincipals {
+		if this.AllowedPrincipals[i] != that1.AllowedPrincipals[i] {
+			return false
+		}
+	}
 	return true
 }
 func (this *NamespaceSpec) GoString() string {
@@ -1039,7 +1073,7 @@ func (this *Namespace) GoString() string {
 	if this == nil {
 		return "nil"
 	}
-	s := make([]string, 0, 13)
+	s := make([]string, 0, 14)
 	s = append(s, "&namespace.Namespace{")
 	s = append(s, "Namespace: "+fmt.Sprintf("%#v", this.Namespace)+",\n")
 	s = append(s, "ResourceVersion: "+fmt.Sprintf("%#v", this.ResourceVersion)+",\n")
@@ -1060,6 +1094,7 @@ func (this *Namespace) GoString() string {
 	if this.CodecServerProperty != nil {
 		s = append(s, "CodecServerProperty: "+fmt.Sprintf("%#v", this.CodecServerProperty)+",\n")
 	}
+	s = append(s, "AllowedPrincipals: "+fmt.Sprintf("%#v", this.AllowedPrincipals)+",\n")
 	s = append(s, "}")
 	return strings.Join(s, "")
 }
@@ -1365,6 +1400,15 @@ func (m *Namespace) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
+	if len(m.AllowedPrincipals) > 0 {
+		for iNdEx := len(m.AllowedPrincipals) - 1; iNdEx >= 0; iNdEx-- {
+			i -= len(m.AllowedPrincipals[iNdEx])
+			copy(dAtA[i:], m.AllowedPrincipals[iNdEx])
+			i = encodeVarintMessage(dAtA, i, uint64(len(m.AllowedPrincipals[iNdEx])))
+			i--
+			dAtA[i] = 0x5a
+		}
+	}
 	if m.CodecServerProperty != nil {
 		{
 			size, err := m.CodecServerProperty.MarshalToSizedBuffer(dAtA[:i])
@@ -1632,6 +1676,12 @@ func (m *Namespace) Size() (n int) {
 		l = m.CodecServerProperty.Size()
 		n += 1 + l + sovMessage(uint64(l))
 	}
+	if len(m.AllowedPrincipals) > 0 {
+		for _, s := range m.AllowedPrincipals {
+			l = len(s)
+			n += 1 + l + sovMessage(uint64(l))
+		}
+	}
 	return n
 }
 
@@ -1734,6 +1784,7 @@ func (this *Namespace) String() string {
 		`Uri:` + strings.Replace(this.Uri.String(), "NamespaceURI", "NamespaceURI", 1) + `,`,
 		`Envelope:` + strings.Replace(this.Envelope.String(), "NamespaceEnvelope", "NamespaceEnvelope", 1) + `,`,
 		`CodecServerProperty:` + strings.Replace(this.CodecServerProperty.String(), "CodecServerPropertySpec", "CodecServerPropertySpec", 1) + `,`,
+		`AllowedPrincipals:` + fmt.Sprintf("%v", this.AllowedPrincipals) + `,`,
 		`}`,
 	}, "")
 	return s
@@ -2966,6 +3017,38 @@ func (m *Namespace) Unmarshal(dAtA []byte) error {
 			if err := m.CodecServerProperty.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
+			iNdEx = postIndex
+		case 11:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field AllowedPrincipals", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowMessage
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthMessage
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthMessage
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.AllowedPrincipals = append(m.AllowedPrincipals, string(dAtA[iNdEx:postIndex]))
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
